@@ -1,89 +1,251 @@
 """Topology builder."""
 
+from __future__ import annotations
+
 import asyncio
+import collections
+import itertools
 import logging
-import time
-from typing import Dict, Optional
+import random
+import typing
 
 import zigpy.config
-import zigpy.neighbor
+import zigpy.device
 import zigpy.types as t
-import zigpy.typing
 import zigpy.util
+import zigpy.zdo.types as zdo_t
 
 LOGGER = logging.getLogger(__name__)
-DELAY_INTER_DEVICE = 3
+REQUEST_DELAY = (1.0, 1.5)
+
+if typing.TYPE_CHECKING:
+    import zigpy.application
+
+
+class ScanNotSupported(Exception):
+    pass
+
+
+INVALID_NEIGHBOR_IEEES = {
+    t.EUI64.convert("00:00:00:00:00:00:00:00"),
+    t.EUI64.convert("ff:ff:ff:ff:ff:ff:ff:ff"),
+}
 
 
 class Topology(zigpy.util.ListenableMixin):
     """Topology scanner."""
 
-    def __init__(self, app: zigpy.typing.ControllerApplicationType):
+    def __init__(self, app: zigpy.application.ControllerApplication):
         """Instantiate."""
-        self._app: zigpy.typing.ControllerApplicationType = app
-        self._listeners: Dict = {}
-        self._last_scanned: Optional[t.EUI64] = None
-        delay_minutes = app.config[zigpy.config.CONF_TOPO_SCAN_PERIOD]
-        self._scan_period = delay_minutes * 60
-        self._scan_task: Optional[asyncio.Task] = None
-        self._timestamp: float = 0
+        self._app: zigpy.application.ControllerApplication = app
+        self._listeners: dict = {}
+        self._scan_task: asyncio.Task | None = None
+        self._scan_loop_task: asyncio.Task | None = None
 
-    @property
-    def timestamp(self) -> float:
-        """Return timestamp of successful build."""
-        return self._timestamp
+        # Keep track of devices that do not support scanning
+        self._neighbors_unsupported: set[t.EUI64] = set()
+        self._routes_unsupported: set[t.EUI64] = set()
 
-    @classmethod
-    def new(cls, app: zigpy.typing.ControllerApplicationType) -> "Topology":
-        """Create Topology instance."""
+        self.neighbors: dict[t.EUI64, list[zdo_t.Neighbor]] = collections.defaultdict(
+            list
+        )
+        self.routes: dict[t.EUI64, list[zdo_t.Route]] = collections.defaultdict(list)
 
-        topo = cls(app)
-        if app.config[zigpy.config.CONF_TOPO_SCAN_ENABLED]:
-            asyncio.create_task(topo.scan_loop())
-        return topo
+    def start_periodic_scans(self, period: int | float) -> None:
+        self.stop_periodic_scans()
+        self._scan_loop_task = asyncio.create_task(self._scan_loop(period))
 
-    async def scan_loop(self) -> None:
+    def stop_periodic_scans(self):
+        if self._scan_loop_task is not None:
+            self._scan_loop_task.cancel()
+
+    async def _scan_loop(self, period: int | float) -> None:
         """Delay scan by creating a task."""
 
         while True:
-            await asyncio.sleep(self._scan_period)
-            if not self._scan_task or self._scan_task.done():
-                LOGGER.debug("Starting scheduled neighbor scan")
-                await self.scan()
+            await asyncio.sleep(period)
 
-    async def scan(self) -> None:
+            # Don't run a scheduled scan if a scan is already running
+            if self._scan_task is not None and not self._scan_task.done():
+                continue
+
+            LOGGER.debug("Starting scheduled neighbor scan")
+
+            try:
+                await self.scan()
+            except asyncio.CancelledError:
+                # We explicitly catch a cancellation here to ensure the scan loop will
+                # not be interrupted if a manual scan is initiated
+                LOGGER.debug("Topology scan cancelled")
+            except (Exception, asyncio.CancelledError):
+                LOGGER.warning("Topology scan failed", exc_info=True)
+
+    async def scan(
+        self, devices: typing.Iterable[zigpy.device.Device] | None = None
+    ) -> None:
         """Preempt Topology scan and reschedule."""
 
         if self._scan_task and not self._scan_task.done():
             LOGGER.debug("Cancelling old scanning task")
             self._scan_task.cancel()
 
-        self._scan_task = asyncio.create_task(self._scan())
-        try:
-            await self._scan_task
-        except asyncio.CancelledError:
-            LOGGER.warning("Cancelled topology scanning task")
+        self._scan_task = asyncio.create_task(self._scan(devices))
+        await self._scan_task
 
-    async def _scan(self) -> None:
+    async def _scan_table(
+        self, scan_request: typing.Callable, entries_attr: str
+    ) -> list[typing.Any]:
+        """Scan a device table by sending ZDO requests."""
+
+        index = 0
+        table = []
+
+        while True:
+            status, rsp = await scan_request(index, tries=3, delay=1)
+
+            if status != zdo_t.Status.SUCCESS:
+                raise ScanNotSupported()
+
+            entries = getattr(rsp, entries_attr)
+
+            table.extend(entries)
+            index += len(entries)
+
+            # We intentionally sleep after every request, even the last one, to simplify
+            # delay logic when scanning many devices in quick succession
+            await asyncio.sleep(random.uniform(*REQUEST_DELAY))
+
+            if index >= rsp.Entries or not entries:
+                break
+
+        return table
+
+    async def _scan_neighbors(
+        self, device: zigpy.device.Device
+    ) -> list[zdo_t.Neighbor]:
+        if device.ieee in self._neighbors_unsupported:
+            return []
+
+        LOGGER.debug("Scanning neighbors of %s", device)
+
+        try:
+            table = await self._scan_table(device.zdo.Mgmt_Lqi_req, "NeighborTableList")
+        except ScanNotSupported:
+            table = []
+            self._neighbors_unsupported.add(device.ieee)
+
+        return [n for n in table if n.ieee not in INVALID_NEIGHBOR_IEEES]
+
+    async def _scan_routes(self, device: zigpy.device.Device) -> list[zdo_t.Route]:
+        if device.ieee in self._routes_unsupported:
+            return []
+
+        LOGGER.debug("Scanning routing table of %s", device)
+
+        try:
+            table = await self._scan_table(device.zdo.Mgmt_Rtg_req, "RoutingTableList")
+        except ScanNotSupported:
+            table = []
+            self._routes_unsupported.add(device.ieee)
+
+        return table
+
+    async def _scan(
+        self, devices: typing.Iterable[zigpy.device.Device] | None = None
+    ) -> None:
         """Scan topology."""
 
-        devices_to_scan = [
-            dev
-            for dev in self._app.devices.values()
-            if dev.node_desc is not None and not dev.node_desc.is_end_device
-        ]
-        for device in devices_to_scan:
-            if (
-                self._app.config[zigpy.config.CONF_TOPO_SKIP_COORDINATOR]
-                and device.nwk == 0x0000
+        if devices is None:
+            # We iterate over a copy of the devices as opposed to the live dictionary
+            devices = list(self._app.devices.values())
+
+        for index, device in enumerate(devices):
+            LOGGER.debug(
+                "Scanning topology (%d/%d) of %s", index + 1, len(devices), device
+            )
+
+            # Ignore devices that aren't routers
+            if device.node_desc is None or not (
+                device.node_desc.is_router or device.node_desc.is_coordinator
             ):
                 continue
-            if not device.neighbors.supported:
+
+            # Ignore devices that do not support scanning tables
+            if (
+                device.ieee in self._neighbors_unsupported
+                and device.ieee in self._routes_unsupported
+            ):
                 continue
-            LOGGER.debug(
-                "Scanning neighbors of %s/0x%04x device", device.ieee, device.nwk
+
+            # Some coordinators have issues when performing loopback scans
+            if (
+                self._app.config[zigpy.config.CONF_TOPO_SKIP_COORDINATOR]
+                and device is self._app._device
+            ):
+                continue
+
+            try:
+                self.neighbors[device.ieee] = await self._scan_neighbors(device)
+            except Exception as e:
+                LOGGER.debug("Failed to scan neighbors of %s", device, exc_info=e)
+            else:
+                LOGGER.info(
+                    "Scanned neighbors of %s: %s", device, self.neighbors[device.ieee]
+                )
+
+            self.listener_event(
+                "neighbors_updated", device.ieee, self.neighbors[device.ieee]
             )
-            await device.neighbors.scan()
-            await asyncio.sleep(DELAY_INTER_DEVICE)
+
+            try:
+                # Filter out inactive routes
+                routes = await self._scan_routes(device)
+                self.routes[device.ieee] = [
+                    route
+                    for route in routes
+                    if route.RouteStatus != zdo_t.RouteStatus.Inactive
+                ]
+            except Exception as e:
+                LOGGER.debug("Failed to scan routes of %s", device, exc_info=e)
+            else:
+                LOGGER.info(
+                    "Scanned routes of %s: %s", device, self.routes[device.ieee]
+                )
+
+            self.listener_event("routes_updated", device.ieee, self.routes[device.ieee])
+
         LOGGER.debug("Finished scanning neighbors for all devices")
-        self._timestamp = time.time()
+        await self._find_unknown_devices(neighbors=self.neighbors, routes=self.routes)
+
+    async def _find_unknown_devices(
+        self,
+        *,
+        neighbors: dict[t.EUI64, list[zdo_t.Neighbor]],
+        routes: dict[t.EUI64, list[zdo_t.Route]],
+    ) -> None:
+        """Discover unknown devices discovered during topology scanning"""
+        # Build a list of unknown devices from the topology scan
+        unknown_nwks = set()
+
+        for neighbor in itertools.chain.from_iterable(neighbors.values()):
+            try:
+                self._app.get_device(nwk=neighbor.nwk)
+            except KeyError:
+                unknown_nwks.add(neighbor.nwk)
+
+        for route in itertools.chain.from_iterable(routes.values()):
+            # Ignore inactive or pending routes
+            if route.RouteStatus != zdo_t.RouteStatus.Active:
+                continue
+
+            for nwk in (route.DstNWK, route.NextHop):
+                try:
+                    self._app.get_device(nwk=nwk)
+                except KeyError:
+                    unknown_nwks.add(nwk)
+
+        # Try to discover any unknown devices
+        for nwk in unknown_nwks:
+            LOGGER.debug("Found unknown device nwk=%s", nwk)
+            await self._app._discover_unknown_device(nwk)
+            await asyncio.sleep(random.uniform(*REQUEST_DELAY))
