@@ -14,44 +14,24 @@ from zigpy.exceptions import (
 )
 import zigpy.ota
 import zigpy.quirks
-import zigpy.state as app_state
 import zigpy.types as t
 import zigpy.zdo.types as zdo_t
 
 from .async_mock import AsyncMock, MagicMock, patch, sentinel
-from .conftest import App
-
-NCP_IEEE = t.EUI64.convert("aa:11:22:bb:33:44:be:ef")
-
-
-@pytest.fixture
-@patch("zigpy.ota.OTA", MagicMock(spec_set=zigpy.ota.OTA))
-@patch("zigpy.device.Device._initialize", AsyncMock())
-def app_factory():
-    def app(extra_config={}):
-        config = {
-            conf.CONF_DATABASE: None,
-            conf.CONF_DEVICE: {conf.CONF_DEVICE_PATH: "/dev/null"},
-        }
-        config.update(extra_config)
-
-        app = App(config)
-        app.state.node_info = app_state.NodeInfo(
-            nwk=t.NWK(0x0000), ieee=NCP_IEEE, logical_type=zdo_t.LogicalType.Coordinator
-        )
-        return app
-
-    return app
+from .conftest import (
+    NCP_IEEE,
+    App,
+    make_app,
+    make_ieee,
+    make_neighbor,
+    make_neighbor_from_device,
+    make_node_desc,
+)
 
 
 @pytest.fixture
-def app(app_factory):
-    return app_factory()
-
-
-@pytest.fixture
-def ieee(init=0):
-    return t.EUI64(map(t.uint8_t, range(init, init + 8)))
+def ieee():
+    return make_ieee()
 
 
 @patch("zigpy.ota.OTA", spec_set=zigpy.ota.OTA)
@@ -110,10 +90,8 @@ async def test_permit(app, ieee):
 
 
 async def test_permit_delivery_failure(app, ieee):
-    from zigpy.exceptions import DeliveryError
-
     def zdo_permit(*args, **kwargs):
-        raise DeliveryError
+        raise DeliveryError("Failed")
 
     app.devices[ieee] = MagicMock()
     app.devices[ieee].zdo.permit = zdo_permit
@@ -123,20 +101,24 @@ async def test_permit_delivery_failure(app, ieee):
 
 
 async def test_permit_broadcast(app):
-    app.broadcast = AsyncMock()
     app.permit_ncp = AsyncMock()
+    app.send_packet = AsyncMock()
     await app.permit(time_s=30)
-    assert app.broadcast.call_count == 1
+    assert app.send_packet.call_count == 1
     assert app.permit_ncp.call_count == 1
+
+    assert app.send_packet.mock_calls[0].args[0].dst.addr_mode == t.AddrMode.Broadcast
 
 
 @patch("zigpy.device.Device.initialize", new_callable=AsyncMock)
 async def test_join_handler_skip(init_mock, app, ieee):
-    app.handle_join(1, ieee, None)
-    app.get_device(ieee).node_desc = _devices(1).node_desc
+    node_desc = make_node_desc()
 
     app.handle_join(1, ieee, None)
-    assert app.get_device(ieee).node_desc == _devices(1).node_desc
+    app.get_device(ieee).node_desc = node_desc
+
+    app.handle_join(1, ieee, None)
+    assert app.get_device(ieee).node_desc == node_desc
 
 
 async def test_join_handler_change_id(app, ieee):
@@ -166,7 +148,7 @@ async def _remove(
         if zdo_reply:
             return retval
         elif delivery_failure:
-            raise DeliveryError
+            raise DeliveryError("Error")
         else:
             raise asyncio.TimeoutError
 
@@ -396,49 +378,34 @@ async def test_uninitialized_message_handlers(app, ieee):
     assert handler_2.call_count == 1
 
 
-def _devices(index):
-    """Device factory."""
-
-    start_ieee = 0xFEAB000000
-    start_nwk = 0x1000
-
-    dev = MagicMock()
-    dev.ieee = zigpy.types.EUI64(zigpy.types.uint64_t(start_ieee + index).serialize())
-    dev.nwk = zigpy.types.NWK(start_nwk + index)
-    dev.neighbors = []
-    dev.node_desc = zdo_t.NodeDescriptor(1, 64, 142, 4388, 82, 255, 0, 255, 0)
-    dev.zdo = zigpy.zdo.ZDO(dev)
-    return dev
-
-
-async def test_remove_parent_devices(app):
+async def test_remove_parent_devices(app, make_initialized_device):
     """Test removing an end device with parents."""
 
-    end_device = _devices(1)
+    end_device = make_initialized_device(app)
     end_device.node_desc.logical_type = zdo_t.LogicalType.EndDevice
-    nei_end_device = MagicMock()
-    nei_end_device.device = end_device
 
-    router_1 = _devices(2)
-    nei_router_1 = MagicMock()
-    nei_router_1.device = router_1
+    router_1 = make_initialized_device(app)
+    router_1.node_desc.logical_type = zdo_t.LogicalType.Router
 
-    router_2 = _devices(3)
-    nei_router_2 = MagicMock()
-    nei_router_2.device = router_2
+    router_2 = make_initialized_device(app)
+    router_2.node_desc.logical_type = zdo_t.LogicalType.Router
 
-    parent = _devices(4)
-    nei_parent = MagicMock()
-    nei_parent.device = router_1
+    parent = make_initialized_device(app)
 
-    router_1.neighbors = [nei_router_2, nei_parent]
-    router_2.neighbors = [nei_parent, nei_router_1]
-    parent.neighbors = [nei_router_2, nei_router_1, nei_end_device]
-
-    app.devices[end_device.ieee] = end_device
-    app.devices[parent.ieee] = parent
-    app.devices[router_1.ieee] = router_1
-    app.devices[router_2.ieee] = router_2
+    app.topology.neighbors[router_1.ieee] = [
+        make_neighbor_from_device(router_2),
+        make_neighbor_from_device(parent),
+    ]
+    app.topology.neighbors[router_2.ieee] = [
+        make_neighbor_from_device(parent),
+        make_neighbor_from_device(router_1),
+    ]
+    app.topology.neighbors[parent.ieee] = [
+        make_neighbor_from_device(router_2),
+        make_neighbor_from_device(router_1),
+        make_neighbor_from_device(end_device),
+        make_neighbor(ieee=make_ieee(123), nwk=0x9876),
+    ]
 
     p1 = patch.object(end_device.zdo, "leave", AsyncMock())
     p2 = patch.object(end_device.zdo, "request", AsyncMock())
@@ -683,8 +650,8 @@ async def test_deprecated_properties_and_methods(app):
         assert app.nwk_update_id is app.state.network_info.nwk_update_id
 
 
-async def test_startup_backup(app_factory):
-    app = app_factory({conf.CONF_NWK_BACKUP_ENABLED: True})
+async def test_startup_backup():
+    app = make_app({conf.CONF_NWK_BACKUP_ENABLED: True})
 
     with patch("zigpy.backups.BackupManager.start_periodic_backups") as p:
         await app.startup()
@@ -692,8 +659,8 @@ async def test_startup_backup(app_factory):
     p.assert_called_once()
 
 
-async def test_startup_no_backup(app_factory):
-    app = app_factory({conf.CONF_NWK_BACKUP_ENABLED: False})
+async def test_startup_no_backup():
+    app = make_app({conf.CONF_NWK_BACKUP_ENABLED: False})
 
     with patch("zigpy.backups.BackupManager.start_periodic_backups") as p:
         await app.startup()
@@ -704,9 +671,9 @@ async def test_startup_no_backup(app_factory):
 @patch("zigpy.backups.BackupManager.from_network_state")
 @patch("zigpy.backups.BackupManager.most_recent_backup")
 async def test_initialize_compatible_backup(
-    mock_most_recent_backup, mock_backup_from_state, app_factory
+    mock_most_recent_backup, mock_backup_from_state
 ):
-    app = app_factory({conf.CONF_NWK_VALIDATE_SETTINGS: True})
+    app = make_app({conf.CONF_NWK_VALIDATE_SETTINGS: True})
     mock_backup_from_state.return_value.is_compatible_with.return_value = True
 
     await app.initialize()
@@ -718,9 +685,9 @@ async def test_initialize_compatible_backup(
 @patch("zigpy.backups.BackupManager.from_network_state")
 @patch("zigpy.backups.BackupManager.most_recent_backup")
 async def test_initialize_incompatible_backup(
-    mock_most_recent_backup, mock_backup_from_state, app_factory
+    mock_most_recent_backup, mock_backup_from_state
 ):
-    app = app_factory({conf.CONF_NWK_VALIDATE_SETTINGS: True})
+    app = make_app({conf.CONF_NWK_VALIDATE_SETTINGS: True})
     mock_backup_from_state.return_value.is_compatible_with.return_value = False
 
     with pytest.raises(NetworkSettingsInconsistent):
@@ -728,3 +695,366 @@ async def test_initialize_incompatible_backup(
 
     mock_backup_from_state.return_value.is_compatible_with.assert_called_once()
     mock_most_recent_backup.assert_called_once()
+
+
+async def test_relays_received_device_exists(app):
+    device = MagicMock()
+
+    app._discover_unknown_device = AsyncMock(spec_set=app._discover_unknown_device)
+    app.get_device = MagicMock(spec_set=app.get_device, return_value=device)
+    app.handle_relays(nwk=0x1234, relays=[0x5678, 0xABCD])
+
+    app.get_device.assert_called_once_with(nwk=0x1234)
+    assert device.relays == [0x5678, 0xABCD]
+    assert app._discover_unknown_device.call_count == 0
+
+
+async def test_relays_received_device_does_not_exist(app):
+    app._discover_unknown_device = AsyncMock(spec_set=app._discover_unknown_device)
+    app.get_device = MagicMock(wraps=app.get_device)
+    app.handle_relays(nwk=0x1234, relays=[0x5678, 0xABCD])
+
+    app.get_device.assert_called_once_with(nwk=0x1234)
+    app._discover_unknown_device.assert_called_once_with(nwk=0x1234)
+
+
+async def test_request_concurrency():
+    current_concurrency = 0
+    peak_concurrency = 0
+
+    class SlowApp(App):
+        async def send_packet(self, packet):
+            nonlocal current_concurrency, peak_concurrency
+
+            async with self._limit_concurrency():
+                current_concurrency += 1
+                peak_concurrency = max(peak_concurrency, current_concurrency)
+
+                await asyncio.sleep(0.1)
+                current_concurrency -= 1
+
+                if packet % 10 == 7:
+                    # Fail randomly
+                    raise asyncio.DeliveryError()
+
+    app = make_app({conf.CONF_MAX_CONCURRENT_REQUESTS: 16}, app_base=SlowApp)
+
+    assert current_concurrency == 0
+    assert peak_concurrency == 0
+
+    await asyncio.gather(
+        *[app.send_packet(i) for i in range(100)], return_exceptions=True
+    )
+
+    assert current_concurrency == 0
+    assert peak_concurrency == 16
+
+
+@pytest.fixture
+def device():
+    device = MagicMock()
+    device.nwk = 0xABCD
+    device.ieee = t.EUI64.convert("aa:bb:cc:dd:11:22:33:44")
+
+    return device
+
+
+@pytest.fixture
+def packet(app, device):
+    return t.ZigbeePacket(
+        src=t.AddrModeAddress(
+            addr_mode=t.AddrMode.NWK, address=app.state.node_info.nwk
+        ),
+        src_ep=0x9A,
+        dst=t.AddrModeAddress(addr_mode=t.AddrMode.NWK, address=device.nwk),
+        dst_ep=0xBC,
+        tsn=0xDE,
+        profile_id=0x1234,
+        cluster_id=0x0006,
+        data=t.SerializableBytes(b"test data"),
+        source_route=None,
+        extended_timeout=False,
+        tx_options=t.TransmitOptions.NONE,
+    )
+
+
+async def test_request(app, device, packet):
+    app.build_source_route_to = MagicMock(spec_set=app.build_source_route_to)
+
+    async def send_request(app, **kwargs):
+        kwargs = {
+            "device": device,
+            "profile": 0x1234,
+            "cluster": 0x0006,
+            "src_ep": 0x9A,
+            "dst_ep": 0xBC,
+            "sequence": 0xDE,
+            "data": b"test data",
+            "expect_reply": True,
+            "use_ieee": False,
+            "extended_timeout": False,
+            **kwargs,
+        }
+
+        return await app.request(**kwargs)
+
+    # Test sending with NWK
+    status, msg = await send_request(app)
+    assert status == zigpy.zcl.foundation.Status.SUCCESS
+    assert isinstance(msg, str)
+
+    app.send_packet.assert_called_once_with(packet)
+    app.send_packet.reset_mock()
+
+    # Test sending with IEEE
+    await send_request(app, use_ieee=True)
+    app.send_packet.assert_called_once_with(
+        packet.replace(
+            src=t.AddrModeAddress(
+                addr_mode=t.AddrMode.IEEE, address=app.state.node_info.ieee
+            ),
+            dst=t.AddrModeAddress(addr_mode=t.AddrMode.IEEE, address=device.ieee),
+        )
+    )
+    app.send_packet.reset_mock()
+
+    # Test sending with source route
+    app.build_source_route_to.return_value = [0x000A, 0x000B]
+
+    with patch.dict(app.config, {conf.CONF_SOURCE_ROUTING: True}):
+        await send_request(app)
+
+    app.build_source_route_to.assert_called_once_with(dest=device)
+    app.send_packet.assert_called_once_with(
+        packet.replace(source_route=[0x000A, 0x000B])
+    )
+    app.send_packet.reset_mock()
+
+    # Test sending without waiting for a reply
+    status, msg = await send_request(app, expect_reply=False)
+
+    app.send_packet.assert_called_once_with(
+        packet.replace(tx_options=t.TransmitOptions.ACK)
+    )
+    app.send_packet.reset_mock()
+
+
+def test_build_source_route_has_relays(app):
+    device = MagicMock()
+    device.relays = [0x1234, 0x5678]
+
+    assert app.build_source_route_to(device) == [0x5678, 0x1234]
+
+
+def test_build_source_route_no_relays(app):
+    device = MagicMock()
+    device.relays = None
+
+    assert app.build_source_route_to(device) is None
+
+
+async def test_send_mrequest(app, packet):
+
+    status, msg = await app.mrequest(
+        group_id=0xABCD,
+        profile=0x1234,
+        cluster=0x0006,
+        src_ep=0x9A,
+        sequence=0xDE,
+        data=b"test data",
+        hops=12,
+        non_member_radius=34,
+    )
+    assert status == zigpy.zcl.foundation.Status.SUCCESS
+    assert isinstance(msg, str)
+
+    app.send_packet.assert_called_once_with(
+        packet.replace(
+            dst=t.AddrModeAddress(addr_mode=t.AddrMode.Group, address=0xABCD),
+            dst_ep=None,
+            radius=12,
+            non_member_radius=34,
+            tx_options=t.TransmitOptions.NONE,
+        )
+    )
+
+
+async def test_send_broadcast(app, packet):
+
+    status, msg = await app.broadcast(
+        profile=0x1234,
+        cluster=0x0006,
+        src_ep=0x9A,
+        dst_ep=0xBC,
+        grpid=0x0000,  # unused
+        radius=12,
+        sequence=0xDE,
+        data=b"test data",
+        broadcast_address=t.BroadcastAddress.RX_ON_WHEN_IDLE,
+    )
+    assert status == zigpy.zcl.foundation.Status.SUCCESS
+    assert isinstance(msg, str)
+
+    app.send_packet.assert_called_once_with(
+        packet.replace(
+            dst=t.AddrModeAddress(
+                addr_mode=t.AddrMode.Broadcast,
+                address=t.BroadcastAddress.RX_ON_WHEN_IDLE,
+            ),
+            radius=12,
+            tx_options=t.TransmitOptions.NONE,
+        )
+    )
+
+
+@pytest.fixture
+def zdo_packet(app, device):
+    return t.ZigbeePacket(
+        src=t.AddrModeAddress(addr_mode=t.AddrMode.NWK, address=device.nwk),
+        dst=t.AddrModeAddress(
+            addr_mode=t.AddrMode.NWK, address=app.state.node_info.nwk
+        ),
+        src_ep=0x00,  # ZDO
+        dst_ep=0x00,
+        tsn=0xDE,
+        profile_id=0x0000,
+        cluster_id=0x0000,
+        data=t.SerializableBytes(b""),
+        source_route=None,
+        extended_timeout=False,
+        tx_options=t.TransmitOptions.ACK,
+        lqi=123,
+        rssi=-80,
+    )
+
+
+@patch("zigpy.device.Device.initialize", AsyncMock())
+async def test_packet_received_new_device_zdo_announce(app, device, zdo_packet):
+    app.handle_join = MagicMock(wraps=app.handle_join)
+
+    zdo_data = zigpy.zdo.ZDO(None)._serialize(
+        zdo_t.ZDOCmd.Device_annce,
+        *dict(
+            NWKAddr=device.nwk,
+            IEEEAddr=device.ieee,
+            Capability=0x00,
+        ).values()
+    )
+
+    zdo_packet.cluster_id = zdo_t.ZDOCmd.Device_annce
+    zdo_packet.data = t.SerializableBytes(
+        t.uint8_t(zdo_packet.tsn).serialize() + zdo_data
+    )
+    app.packet_received(zdo_packet)
+
+    app.handle_join.assert_called_once_with(
+        nwk=device.nwk, ieee=device.ieee, parent_nwk=None
+    )
+
+    zigpy_device = app.get_device(ieee=device.ieee)
+    assert zigpy_device.lqi == zdo_packet.lqi
+    assert zigpy_device.rssi == zdo_packet.rssi
+
+
+@patch("zigpy.device.Device.initialize", AsyncMock())
+async def test_packet_received_new_device_discovery(app, device, zdo_packet):
+    app.handle_join = MagicMock(wraps=app.handle_join)
+
+    async def send_packet(packet):
+        if packet.dst_ep != 0x00 or packet.cluster_id != zdo_t.ZDOCmd.IEEE_addr_req:
+            return
+
+        hdr, args = zigpy.zdo.ZDO(None).deserialize(
+            packet.cluster_id, packet.data.serialize()
+        )
+        assert args == list(
+            dict(
+                NWKAddrOfInterest=device.nwk,
+                RequestType=zdo_t.AddrRequestType.Single,
+                StartIndex=0,
+            ).values()
+        )
+
+        zdo_data = zigpy.zdo.ZDO(None)._serialize(
+            zdo_t.ZDOCmd.IEEE_addr_rsp,
+            *dict(
+                Status=zdo_t.Status.SUCCESS,
+                IEEEAddr=device.ieee,
+                NWKAddr=device.nwk,
+                NumAssocDev=0,
+                StartIndex=0,
+                NWKAddrAssocDevList=[],
+            ).values()
+        )
+
+        # Receive the IEEE address reply
+        zdo_packet.data = t.SerializableBytes(
+            t.uint8_t(zdo_packet.tsn).serialize() + zdo_data
+        )
+        zdo_packet.cluster_id = zdo_t.ZDOCmd.IEEE_addr_rsp
+        app.packet_received(zdo_packet)
+
+    app.send_packet = AsyncMock(side_effect=send_packet)
+
+    # Receive a bogus packet first, to trigger device discovery
+    bogus_packet = zdo_packet.replace(dst_ep=0x01, src_ep=0x01)
+    app.packet_received(bogus_packet)
+
+    await asyncio.sleep(0.1)
+
+    app.handle_join.assert_called_once_with(
+        nwk=device.nwk, ieee=device.ieee, parent_nwk=None
+    )
+
+    zigpy_device = app.get_device(ieee=device.ieee)
+    assert zigpy_device.lqi == zdo_packet.lqi
+    assert zigpy_device.rssi == zdo_packet.rssi
+
+
+async def test_bad_zdo_packet_received(app, device):
+    device.is_initialized = True
+    app.devices[device.ieee] = device
+
+    bogus_zdo_packet = t.ZigbeePacket(
+        src=t.AddrModeAddress(addr_mode=t.AddrMode.NWK, address=device.nwk),
+        src_ep=1,
+        dst=t.AddrModeAddress(addr_mode=t.AddrMode.NWK, address=0x0000),
+        dst_ep=0,  # bad destination endpoint
+        tsn=180,
+        profile_id=260,
+        cluster_id=6,
+        data=t.SerializableBytes(b"\x08n\n\x00\x00\x10\x00"),
+        lqi=255,
+        rssi=-30,
+    )
+
+    app.packet_received(bogus_zdo_packet)
+
+    assert len(device.handle_message.mock_calls) == 1
+
+
+def test_get_device_with_address_nwk(app, device):
+    app.devices[device.ieee] = device
+
+    assert (
+        app.get_device_with_address(
+            t.AddrModeAddress(addr_mode=t.AddrMode.NWK, address=device.nwk)
+        )
+        is device
+    )
+    assert (
+        app.get_device_with_address(
+            t.AddrModeAddress(addr_mode=t.AddrMode.IEEE, address=device.ieee)
+        )
+        is device
+    )
+
+    with pytest.raises(ValueError):
+        app.get_device_with_address(
+            t.AddrModeAddress(addr_mode=t.AddrMode.Group, address=device.nwk)
+        )
+
+    with pytest.raises(KeyError):
+        app.get_device_with_address(
+            t.AddrModeAddress(addr_mode=t.AddrMode.NWK, address=device.nwk + 1)
+        )
