@@ -12,6 +12,8 @@ def test_abstract_ints():
     assert not issubclass(t.uint8_t, t.int_t)
     assert t.int_t._signed is True
     assert t.uint_t._signed is False
+    assert t.int_t._byteorder == "little"
+    assert t.int_t_be._byteorder == "big"
 
     with pytest.raises(TypeError):
         t.int_t(0)
@@ -45,6 +47,8 @@ def test_int_too_short():
 def test_fractional_ints_corner():
     assert t.uint1_t._size is None
     assert t.uint1_t._bits == 1
+    assert t.uint1_t.min_value == 0
+    assert t.uint1_t.max_value == 1
 
     assert t.uint1_t(0) == 0
     assert t.uint1_t(1) == 1
@@ -70,6 +74,8 @@ def test_fractional_ints_corner():
 def test_fractional_ints_larger():
     assert t.uint7_t._size is None
     assert t.uint7_t._bits == 7
+    assert t.uint7_t.min_value == 0
+    assert t.uint7_t.max_value == 2**7 - 1
 
     assert t.uint7_t(0) == 0
     assert t.uint7_t(1) == 1
@@ -135,6 +141,18 @@ def test_ints_signed():
     n = t.int8s(-126)
     bits = [1, 0] + t.Bits.deserialize(n.serialize())[0]
     assert t.int8s.from_bits(bits) == (n, [1, 0])
+
+
+def test_bigendian_ints():
+    assert t.uint32_t_be(0x12345678).serialize() == b"\x12\x34\x56\x78"
+    assert t.uint32_t_be.deserialize(b"\x12\x34\x56\x78") == (0x12345678, b"")
+    assert t.int32s_be(0x12345678).serialize() == b"\x12\x34\x56\x78"
+    assert t.int32s_be(-1).serialize() == b"\xff\xff\xff\xff"
+    assert t.int32s_be.deserialize(b"\xfe\xdc\xba\x98") == (-0x01234568, b"")
+    assert (
+        t.uint32_t_be(0x12345678).serialize()[::-1]
+        == t.uint32_t(0x12345678).serialize()
+    )
 
 
 def test_bits():
@@ -778,63 +796,26 @@ def test_bitmap_instance_types():
     assert isinstance(TestBitmap(0xFF00), TestBitmap)
 
 
-def test_addressing_ieee():
-    """Test addressing mode ieee."""
-
-    data = b"\x03\x08\x07\x06\x05\x04\x03\x02\x01\x20"
-    extra = b"extra data"
-
-    r, rest = t.Addressing.deserialize(data + extra)
-    assert rest == extra
-    assert r.addr_mode == t.AddrMode.IEEE
-    assert r.addr == t.EUI64.convert("01:02:03:04:05:06:07:08")
-    assert r.endpoint == 32
-
-    assert r.serialize() == data
-    dst = t.Addressing.ieee([8, 7, 6, 5, 4, 3, 2, 1], 32)
-    assert dst.serialize() == data
-
-
-def test_addressing_group():
-    """Test addressing mode group."""
-
-    data = b"\x01\x08\x07"
-    extra = b"extra data"
-
-    r, rest = t.Addressing.deserialize(data + extra)
-    assert rest == extra
-    assert r.addr_mode == t.AddrMode.Group
-    assert r.addr == t.Group(0x0708)
-
-    assert r.serialize() == data
-    dst = t.Addressing.group(0x0708)
-    assert dst.serialize() == data
-
-
-def test_addressing_nwk():
-    """Test addressing mode nwk."""
-
-    data = b"\x02\x07\x08\x2A"
-    extra = b"extra data"
-
-    r, rest = t.Addressing.deserialize(data + extra)
-    assert rest == extra
-    assert r.addr_mode == t.AddrMode.NWK
-    assert r.addr == t.NWK(0x0807)
-
-    assert r.serialize() == data
-    dst = t.Addressing.nwk(0x0807, 42)
-    assert dst.serialize() == data
-
-
-def test_invalid_addressing():
-    """Invalid addressing mode."""
-
-    with pytest.raises(ValueError):
-        t.Addressing.deserialize(b"\x05invalid")
-
-
 def test_nwk_convert():
     assert t.NWK.convert(str(t.NWK(0x1234))[2:]) == t.NWK(0x1234)
     assert str(t.NWK(0x0012))[2:] == "0012"
     assert str(t.NWK(0x1200))[2:] == "1200"
+
+
+def test_serializable_bytes():
+    obj = t.SerializableBytes(b"test")
+    assert obj == obj
+    assert obj == t.SerializableBytes(b"test")
+    assert t.SerializableBytes(obj) == obj
+    assert obj != b"test"
+    assert obj.serialize() == b"test"
+    assert "test" in repr([obj])
+
+    with pytest.raises(TypeError):
+        obj + b"test"
+
+    with pytest.raises(ValueError):
+        t.SerializableBytes("test")
+
+    with pytest.raises(ValueError):
+        t.SerializableBytes([1, 2, 3])
