@@ -9,8 +9,6 @@ import zigpy.zdo.types as zdo_types
 
 from .async_mock import AsyncMock, MagicMock, patch, sentinel
 
-DEFAULT_SEQUENCE = 123
-
 
 def test_commands():
     for cmdid, cmdspec in zdo.types.CLUSTERS.items():
@@ -23,23 +21,13 @@ def test_commands():
 
 
 @pytest.fixture
-def app():
-    app = MagicMock()
-    app.ieee = t.EUI64(map(t.uint8_t, [8, 9, 10, 11, 12, 13, 14, 15]))
-    app.get_sequence.return_value = DEFAULT_SEQUENCE
-    dst_addr = zdo_types.MultiAddress()
-    dst_addr.addrmode = 3
-    dst_addr.ieee = app.ieee
-    dst_addr.endpoint = 1
-    app.get_dst_address.return_value = dst_addr
-    return app
-
-
-@pytest.fixture
 def zdo_f(app):
     ieee = t.EUI64(map(t.uint8_t, [0, 1, 2, 3, 4, 5, 6, 7]))
     dev = zigpy.device.Device(app, ieee, 65535)
     dev.request = AsyncMock()
+
+    app.devices[dev.ieee] = dev
+
     return zdo.ZDO(dev)
 
 
@@ -51,16 +39,15 @@ def test_deserialize(zdo_f):
 
 
 def test_deserialize_unknown(zdo_f):
-    hdr, args = zdo_f.deserialize(0x0100, b"\x01")
-    assert hdr.tsn == 1
-    assert hdr.is_reply is False
+    with pytest.raises(ValueError):
+        hdr, args = zdo_f.deserialize(0x0100, b"\x01")
 
 
 async def test_request(zdo_f):
     await zdo_f.request(2, 65535)
-    app_mock = zdo_f._device._application
+    app = zdo_f._device._application
     assert zdo_f.device.request.call_count == 1
-    assert app_mock.get_sequence.call_count == 1
+    assert app.get_sequence.call_count == 1
 
 
 async def test_bind(zdo_f):
@@ -106,12 +93,14 @@ async def test_permit(zdo_f):
     assert zdo_f.device.request.call_args[0][1] == 0x0036
 
 
-def test_broadcast(app):
-    zigpy.device.broadcast = MagicMock()
-    zigpy.zdo.broadcast(app, 0x0036, 0, 0, 60, 0)
+async def test_broadcast(app):
+    await zigpy.zdo.broadcast(app, 0x0036, 0, 0, 60, 0)
 
-    assert zigpy.device.broadcast.call_count == 1
-    assert zigpy.device.broadcast.call_args[0][2] == 0x0036
+    assert app.send_packet.call_count == 1
+
+    packet = app.send_packet.mock_calls[0].args[0]
+    assert packet.dst.addr_mode == t.AddrMode.Broadcast
+    assert packet.cluster_id == 0x0036
 
 
 def _handle_match_desc(zdo_f, profile):
@@ -137,7 +126,7 @@ async def test_handle_match_desc_generic(zdo_f):
 
 
 async def test_handle_nwk_addr(zdo_f):
-    ieee = zdo_f._device.application.ieee
+    ieee = zdo_f._device.application.state.node_info.ieee
     zdo_f.reply = MagicMock()
     hdr = MagicMock()
     hdr.command_id = zdo_types.ZDOCmd.NWK_addr_req
@@ -146,7 +135,7 @@ async def test_handle_nwk_addr(zdo_f):
 
 
 async def test_handle_ieee_addr(zdo_f):
-    nwk = zdo_f._device.application.nwk
+    nwk = zdo_f._device.application.state.node_info.nwk
     zdo_f.reply = MagicMock()
     hdr = MagicMock()
     hdr.command_id = zdo_types.ZDOCmd.IEEE_addr_req
@@ -204,9 +193,9 @@ def test_device_accessor(zdo_f):
     assert zdo_f.device.nwk == 65535
 
 
-def test_reply(zdo_f):
-    zdo_f.device.request = MagicMock()
-    zdo_f.reply(0x0005)
+async def test_reply(zdo_f):
+    zdo_f.device.request = AsyncMock()
+    await zdo_f.reply(0x0005)
     assert zdo_f.device.request.call_count == 1
 
 
@@ -228,8 +217,8 @@ async def test_reply_tsn_override(zdo_f, monkeypatch):
     await zdo_f.reply(sentinel.cmd, sentinel.arg1, sentinel.arg2)
     seq = zdo_f.device.request.call_args[0][4]
     data = zdo_f.device.request.call_args[0][5]
-    assert seq == DEFAULT_SEQUENCE
-    assert data[0] == DEFAULT_SEQUENCE
+    assert seq == 123
+    assert data[0] == 123
     assert data[1:3] == b"\xaa\x55"
 
     # override tsn
