@@ -1,5 +1,5 @@
 MAJOR_VERSION = 0
-MINOR_VERSION = 51
+MINOR_VERSION = 53
 PATCH_VERSION = "0.dev0"
 __short_version__ = f"{MAJOR_VERSION}.{MINOR_VERSION}"
 __version__ = f"{__short_version__}.{PATCH_VERSION}"
