@@ -1,9 +1,13 @@
-import importlib.resources
+import sys
+
+if sys.version_info >= (3, 9):
+    import importlib.resources as importlib_resources
+else:
+    import importlib_resources
 
 # Map each schema version to its SQL
 SCHEMAS = {}
 
-for resource in importlib.resources.contents(__name__):
-    if resource.startswith("schema_v") and resource.endswith(".sql"):
-        n = int(resource.replace("schema_v", "").replace(".sql", ""), 10)
-        SCHEMAS[n] = importlib.resources.read_text(__name__, resource)
+for file in importlib_resources.files(__name__).glob("schema_v*.sql"):
+    n = int(file.name.replace("schema_v", "").replace(".sql", ""), 10)
+    SCHEMAS[n] = file.read_text()
