@@ -1,5 +1,6 @@
 import zigpy.types as t
 
+CONF_MAX_CONCURRENT_REQUESTS_DEFAULT = 8
 CONF_NWK_BACKUP_ENABLED_DEFAULT = True
 CONF_NWK_BACKUP_PERIOD_DEFAULT = 24 * 60  # 24 hours
 CONF_NWK_CHANNEL_DEFAULT = 15
@@ -15,8 +16,11 @@ CONF_NWK_VALIDATE_SETTINGS_DEFAULT = False
 CONF_OTA_IKEA_DEFAULT = False
 CONF_OTA_INOVELLI_DEFAULT = False
 CONF_OTA_LEDVANCE_DEFAULT = False
-CONF_OTA_SALUS_DEFAULT = False
 CONF_OTA_OTAU_DIR_DEFAULT = None
+CONF_OTA_SALUS_DEFAULT = False
+CONF_OTA_SONOFF_DEFAULT = False
+CONF_OTA_THIRDREALITY_DEFAULT = False
+CONF_SOURCE_ROUTING_DEFAULT = False
 CONF_TOPO_SCAN_PERIOD_DEFAULT = 4 * 60  # 4 hours
 CONF_TOPO_SCAN_ENABLED_DEFAULT = True
 CONF_TOPO_SKIP_COORDINATOR_DEFAULT = False
