@@ -5,7 +5,7 @@ import binascii
 from datetime import datetime, timezone
 import enum
 import logging
-from typing import TYPE_CHECKING, Any
+import typing
 
 from zigpy.const import (
     SIG_ENDPOINTS,
@@ -19,14 +19,14 @@ from zigpy.const import (
 )
 import zigpy.endpoint
 import zigpy.exceptions
-import zigpy.neighbor
-from zigpy.types import NWK, Addressing, BroadcastAddress, Relays
-from zigpy.types.named import EUI64
+import zigpy.types as t
+from zigpy.typing import AddressingMode
 import zigpy.util
 import zigpy.zcl.foundation as foundation
 import zigpy.zdo as zdo
+import zigpy.zdo.types as zdo_t
 
-if TYPE_CHECKING:
+if typing.TYPE_CHECKING:
     from zigpy.application import ControllerApplication
 
 APS_REPLY_TIMEOUT = 5
@@ -50,10 +50,10 @@ class Device(zigpy.util.LocalLogMixin, zigpy.util.ListenableMixin):
 
     manufacturer_id_override = None
 
-    def __init__(self, application, ieee, nwk):
+    def __init__(self, application: ControllerApplication, ieee: t.EUI64, nwk: t.NWK):
         self._application: ControllerApplication = application
-        self._ieee: EUI64 = ieee
-        self.nwk: NWK = NWK(nwk)
+        self._ieee: t.EUI64 = ieee
+        self.nwk: t.NWK = t.NWK(nwk)
         self.zdo: zdo.ZDO = zdo.ZDO(self)
         self.endpoints: dict[int, zdo.ZDO | zigpy.endpoint.Endpoint] = {0: self.zdo}
         self.lqi: int | None = None
@@ -64,10 +64,9 @@ class Device(zigpy.util.LocalLogMixin, zigpy.util.ListenableMixin):
         self._listeners = {}
         self._manufacturer: str | None = None
         self._model: str | None = None
-        self.node_desc: zdo.types.NodeDescriptor | None = None
-        self.neighbors: zigpy.neighbor.Neighbors = zigpy.neighbor.Neighbors(self)
+        self.node_desc: zdo_t.NodeDescriptor | None = None
         self._pending: zigpy.util.Requests = zigpy.util.Requests()
-        self._relays: Relays | None = None
+        self._relays: t.Relays | None = None
         self._skip_configuration: bool = False
 
         # Retained for backwards compatibility, will be removed in a future release
@@ -155,14 +154,14 @@ class Device(zigpy.util.LocalLogMixin, zigpy.util.ListenableMixin):
 
         return self._initialize_task
 
-    async def get_node_descriptor(self) -> zdo.types.NodeDescriptor:
+    async def get_node_descriptor(self) -> zdo_t.NodeDescriptor:
         self.info("Requesting 'Node Descriptor'")
 
         status, _, node_desc = await self.zdo.Node_Desc_req(
             self.nwk, tries=2, delay=0.1
         )
 
-        if status != zdo.types.Status.SUCCESS:
+        if status != zdo_t.Status.SUCCESS:
             raise zigpy.exceptions.InvalidResponse(
                 f"Requesting Node Descriptor failed: {status}"
             )
@@ -180,7 +179,9 @@ class Device(zigpy.util.LocalLogMixin, zigpy.util.ListenableMixin):
                 e, (asyncio.TimeoutError, zigpy.exceptions.ZigbeeException)
             ):
                 LOGGER.warning(
-                    "Device failed to initialize due to unexpected error", exc_info=True
+                    "Device %r failed to initialize due to unexpected error",
+                    self,
+                    exc_info=True,
                 )
 
             self.application.listener_event("device_init_failure", self)
@@ -209,7 +210,7 @@ class Device(zigpy.util.LocalLogMixin, zigpy.util.ListenableMixin):
                 self.nwk, tries=3, delay=0.5
             )
 
-            if status != zdo.types.Status.SUCCESS:
+            if status != zdo_t.Status.SUCCESS:
                 raise zigpy.exceptions.InvalidResponse(
                     f"Endpoint request failed: {status}"
                 )
@@ -284,11 +285,15 @@ class Device(zigpy.util.LocalLogMixin, zigpy.util.ListenableMixin):
         timeout=APS_REPLY_TIMEOUT,
         use_ieee=False,
     ):
+        extended_timeout = False
+
         if expect_reply and (self.node_desc is None or self.node_desc.is_end_device):
             self.debug("Extending timeout for 0x%02x request", sequence)
             timeout = APS_REPLY_TIMEOUT_EXTENDED
+            extended_timeout = True
+
         with self._pending.new(sequence) as req:
-            radio_result, msg = await self._application.request(
+            await self._application.request(
                 self,
                 profile,
                 cluster,
@@ -298,26 +303,8 @@ class Device(zigpy.util.LocalLogMixin, zigpy.util.ListenableMixin):
                 data,
                 expect_reply=expect_reply,
                 use_ieee=use_ieee,
+                extended_timeout=extended_timeout,
             )
-            if radio_result != foundation.Status.SUCCESS:
-                self.debug(
-                    (
-                        "Delivery error for seq # 0x%02x, on endpoint id %s "
-                        "cluster 0x%04x: %s"
-                    ),
-                    sequence,
-                    dst_ep,
-                    cluster,
-                    msg,
-                )
-                raise zigpy.exceptions.DeliveryError(
-                    "[0x{:04x}:{}:0x{:04x}]: Message send failure".format(
-                        self.nwk, dst_ep, cluster
-                    )
-                )
-            # If application.request raises an exception, we won't get here, so
-            # won't update last_seen, as expected
-            self.update_last_seen()
 
             if not expect_reply:
                 return None
@@ -335,8 +322,7 @@ class Device(zigpy.util.LocalLogMixin, zigpy.util.ListenableMixin):
         dst_ep: int,
         message: bytes,
         *,
-        dst_addressing: None
-        | (Addressing.Group | Addressing.IEEE | Addressing.NWK) = None,
+        dst_addressing: AddressingMode | None = None,
     ):
         self.update_last_seen()
 
@@ -385,8 +371,10 @@ class Device(zigpy.util.LocalLogMixin, zigpy.util.ListenableMixin):
             profile, cluster, hdr, args, dst_addressing=dst_addressing
         )
 
-    def reply(self, profile, cluster, src_ep, dst_ep, sequence, data, use_ieee=False):
-        return self.request(
+    async def reply(
+        self, profile, cluster, src_ep, dst_ep, sequence, data, use_ieee=False
+    ):
+        return await self.request(
             profile,
             cluster,
             src_ep,
@@ -411,7 +399,7 @@ class Device(zigpy.util.LocalLogMixin, zigpy.util.ListenableMixin):
         return self._application
 
     @property
-    def ieee(self) -> EUI64:
+    def ieee(self) -> t.EUI64:
         return self._ieee
 
     @property
@@ -454,16 +442,16 @@ class Device(zigpy.util.LocalLogMixin, zigpy.util.ListenableMixin):
             self._skip_configuration = False
 
     @property
-    def relays(self) -> Relays | None:
+    def relays(self) -> t.Relays | None:
         """Relay list."""
         return self._relays
 
     @relays.setter
-    def relays(self, relays: Relays | None) -> None:
+    def relays(self, relays: t.Relays | None) -> None:
         if relays is None:
             pass
-        elif not isinstance(relays, Relays):
-            relays = Relays(relays)
+        elif not isinstance(relays, t.Relays):
+            relays = t.Relays(relays)
 
         self._relays = relays
         self.listener_event("device_relays_updated", relays)
@@ -471,13 +459,13 @@ class Device(zigpy.util.LocalLogMixin, zigpy.util.ListenableMixin):
     def __getitem__(self, key):
         return self.endpoints[key]
 
-    def get_signature(self) -> dict[str, Any]:
+    def get_signature(self) -> dict[str, typing.Any]:
         # return the device signature by providing essential device information
         #    - Model Identifier ( Attribute 0x0005 of Basic Cluster 0x0000 )
         #    - Manufacturer Name ( Attribute 0x0004 of Basic Cluster 0x0000 )
         #    - Endpoint list
         #        - Profile Id, Device Id, Cluster Out, Cluster In
-        signature: dict[str, Any] = {}
+        signature: dict[str, typing.Any] = {}
         if self._manufacturer is not None:
             signature[SIG_MANUFACTURER] = self.manufacturer
         if self._model is not None:
@@ -505,7 +493,7 @@ class Device(zigpy.util.LocalLogMixin, zigpy.util.ListenableMixin):
             f"{type(self).__name__}"
             f" model={self.model!r}"
             f" manuf={self.manufacturer!r}"
-            f" nwk={NWK(self.nwk)}"
+            f" nwk={t.NWK(self.nwk)}"
             f" ieee={self.ieee}"
             f" is_initialized={self.is_initialized}"
             f">"
@@ -522,9 +510,9 @@ async def broadcast(
     radius,
     sequence,
     data,
-    broadcast_address=BroadcastAddress.RX_ON_WHEN_IDLE,
+    broadcast_address=t.BroadcastAddress.RX_ON_WHEN_IDLE,
 ):
-    result = await app.broadcast(
+    return await app.broadcast(
         profile,
         cluster,
         src_ep,
@@ -535,4 +523,3 @@ async def broadcast(
         data,
         broadcast_address=broadcast_address,
     )
-    return result
