@@ -7,9 +7,8 @@ from typing import Any
 
 import zigpy.exceptions
 import zigpy.profiles
-from zigpy.types import Addressing
 from zigpy.types.named import EUI64
-from zigpy.typing import DeviceType
+from zigpy.typing import AddressingMode, DeviceType
 import zigpy.util
 import zigpy.zcl
 from zigpy.zcl.clusters.general import Basic
@@ -211,8 +210,7 @@ class Endpoint(zigpy.util.LocalLogMixin, zigpy.util.ListenableMixin):
         hdr: ZCLHeader,
         args: list,
         *,
-        dst_addressing: None
-        | (Addressing.Group | Addressing.IEEE | Addressing.NWK) = None,
+        dst_addressing: AddressingMode | None = None,
     ) -> None:
         if cluster in self.in_clusters:
             handler = self.in_clusters[cluster].handle_message
@@ -225,7 +223,9 @@ class Endpoint(zigpy.util.LocalLogMixin, zigpy.util.ListenableMixin):
 
         handler(hdr, args, dst_addressing=dst_addressing)
 
-    def request(self, cluster, sequence, data, expect_reply=True, command_id=0x00):
+    async def request(
+        self, cluster, sequence, data, expect_reply=True, command_id=0x00
+    ):
         if self.profile_id == zigpy.profiles.zll.PROFILE_ID and not (
             cluster == zigpy.zcl.clusters.lightlink.LightLink.cluster_id
             and command_id < 0x40
@@ -234,7 +234,7 @@ class Endpoint(zigpy.util.LocalLogMixin, zigpy.util.ListenableMixin):
         else:
             profile_id = self.profile_id
 
-        return self.device.request(
+        return await self.device.request(
             profile_id,
             cluster,
             self._endpoint_id,
@@ -244,7 +244,7 @@ class Endpoint(zigpy.util.LocalLogMixin, zigpy.util.ListenableMixin):
             expect_reply=expect_reply,
         )
 
-    def reply(self, cluster, sequence, data, command_id=0x00):
+    async def reply(self, cluster, sequence, data, command_id=0x00):
         if self.profile_id == zigpy.profiles.zll.PROFILE_ID and not (
             cluster == zigpy.zcl.clusters.lightlink.LightLink.cluster_id
             and command_id < 0x40
@@ -253,7 +253,7 @@ class Endpoint(zigpy.util.LocalLogMixin, zigpy.util.ListenableMixin):
         else:
             profile_id = self.profile_id
 
-        return self.device.reply(
+        return await self.device.reply(
             profile_id, cluster, self._endpoint_id, self._endpoint_id, sequence, data
         )
 
