@@ -3,10 +3,11 @@ from __future__ import annotations
 import enum
 import inspect
 import struct
-from typing import Callable, TypeVar
+import sys
+import typing
 
-CALLABLE_T = TypeVar("CALLABLE_T", bound=Callable)  # pylint: disable=invalid-name
-T = TypeVar("T")
+CALLABLE_T = typing.TypeVar("CALLABLE_T", bound=typing.Callable)
+T = typing.TypeVar("T")
 
 
 class Bits(list):
@@ -47,6 +48,32 @@ class Bits(list):
         return cls(bits), b""
 
 
+class SerializableBytes:
+    """
+    A container object for raw bytes that enforces `serialize()` will be called.
+    """
+
+    def __init__(self, value: bytes = b"") -> None:
+        if isinstance(value, type(self)):
+            value = value.value
+        elif not isinstance(value, (bytes, bytearray)):
+            raise ValueError(f"Object is not bytes: {value!r}")
+
+        self.value = value
+
+    def __eq__(self, other: typing.Any) -> bool:
+        if not isinstance(other, type(self)):
+            return NotImplemented
+
+        return self.value == other.value
+
+    def serialize(self) -> bytes:
+        return self.value
+
+    def __repr__(self) -> str:
+        return f"Serialized[{self.value!r}]"
+
+
 NOT_SET = object()
 
 
@@ -54,6 +81,10 @@ class FixedIntType(int):
     _signed = None
     _bits = None
     _size = None  # Only for backwards compatibility, not set for smaller ints
+    _byteorder = None
+
+    min_value: int
+    max_value: int
 
     def __new__(cls, *args, **kwargs):
         if cls._signed is None or cls._bits is None:
@@ -61,12 +92,11 @@ class FixedIntType(int):
 
         n = super().__new__(cls, *args, **kwargs)
 
-        if not cls._signed and not 0 <= n <= 2**cls._bits - 1:
-            raise ValueError(f"{int(n)} is not an unsigned {cls._bits} bit integer")
-        elif (
-            cls._signed and not -(2 ** (cls._bits - 1)) <= n <= 2 ** (cls._bits - 1) - 1
-        ):
-            raise ValueError(f"{int(n)} is not a signed {cls._bits} bit integer")
+        if not cls.min_value <= n <= cls.max_value:
+            raise ValueError(
+                f"{int(n)} is not an {'un' if not cls._signed else ''}signed"
+                f" {cls._bits} bit integer"
+            )
 
         return n
 
@@ -77,7 +107,9 @@ class FixedIntType(int):
     def _bin_repr(self):
         return f"0b{{:0{self._bits}b}}".format(int(self))
 
-    def __init_subclass__(cls, signed=NOT_SET, bits=NOT_SET, repr=NOT_SET) -> None:
+    def __init_subclass__(
+        cls, signed=NOT_SET, bits=NOT_SET, repr=NOT_SET, byteorder=NOT_SET
+    ) -> None:
         super().__init_subclass__()
 
         if signed is not NOT_SET:
@@ -91,6 +123,14 @@ class FixedIntType(int):
             else:
                 cls._size = None
 
+        if cls._bits is not None and cls._signed is not None:
+            if cls._signed:
+                cls.min_value = -(2 ** (cls._bits - 1))
+                cls.max_value = 2 ** (cls._bits - 1) - 1
+            else:
+                cls.min_value = 0
+                cls.max_value = 2**cls._bits - 1
+
         if repr == "hex":
             assert cls._bits % 4 == 0
             cls.__str__ = cls.__repr__ = cls._hex_repr
@@ -101,6 +141,11 @@ class FixedIntType(int):
             cls.__repr__ = super().__repr__
         elif repr is not NOT_SET:
             raise ValueError(f"Invalid repr value {repr!r}. Must be either hex or bin")
+
+        if byteorder is not NOT_SET:
+            cls._byteorder = byteorder
+        elif cls._byteorder is None:
+            cls._byteorder = "little"
 
         # XXX: The enum module uses the first class with __new__ in its __dict__ as the
         #      member type. We have to ensure this is true for every subclass.
@@ -134,7 +179,7 @@ class FixedIntType(int):
         if self._bits % 8 != 0:
             raise TypeError(f"Integer type with {self._bits} bits is not byte aligned")
 
-        return self.to_bytes(self._bits // 8, "little", signed=self._signed)
+        return self.to_bytes(self._bits // 8, self._byteorder, signed=self._signed)
 
     @classmethod
     def deserialize(cls, data: bytes) -> tuple[FixedIntType, bytes]:
@@ -146,7 +191,7 @@ class FixedIntType(int):
         if len(data) < byte_size:
             raise ValueError(f"Data is too short to contain {byte_size} bytes")
 
-        r = cls.from_bytes(data[:byte_size], "little", signed=cls._signed)
+        r = cls.from_bytes(data[:byte_size], cls._byteorder, signed=cls._signed)
         data = data[byte_size:]
         return r, data
 
@@ -251,6 +296,70 @@ class uint64_t(uint_t, bits=64):
     pass
 
 
+class uint_t_be(FixedIntType, signed=False, byteorder="big"):
+    pass
+
+
+class int_t_be(FixedIntType, signed=True, byteorder="big"):
+    pass
+
+
+class int16s_be(int_t_be, bits=16):
+    pass
+
+
+class int24s_be(int_t_be, bits=24):
+    pass
+
+
+class int32s_be(int_t_be, bits=32):
+    pass
+
+
+class int40s_be(int_t_be, bits=40):
+    pass
+
+
+class int48s_be(int_t_be, bits=48):
+    pass
+
+
+class int56s_be(int_t_be, bits=56):
+    pass
+
+
+class int64s_be(int_t_be, bits=64):
+    pass
+
+
+class uint16_t_be(uint_t_be, bits=16):
+    pass
+
+
+class uint24_t_be(uint_t_be, bits=24):
+    pass
+
+
+class uint32_t_be(uint_t_be, bits=32):
+    pass
+
+
+class uint40_t_be(uint_t_be, bits=40):
+    pass
+
+
+class uint48_t_be(uint_t_be, bits=48):
+    pass
+
+
+class uint56_t_be(uint_t_be, bits=56):
+    pass
+
+
+class uint64_t_be(uint_t_be, bits=64):
+    pass
+
+
 class _IntEnumMeta(enum.EnumMeta):
     def __call__(cls, value, names=None, *args, **kwargs):
         if isinstance(value, str) and value.startswith("0x"):
@@ -266,20 +375,27 @@ def bitmap_factory(int_type: CALLABLE_T) -> CALLABLE_T:
     appropriate methods but with only one non-Enum parent class.
     """
 
-    class _NewEnum(int_type, enum.Flag):
-        # Rebind classmethods to our own class
-        _missing_ = classmethod(enum.IntFlag._missing_.__func__)
-        _create_pseudo_member_ = classmethod(
-            enum.IntFlag._create_pseudo_member_.__func__
-        )
+    if sys.version_info >= (3, 11):
 
-        __or__ = enum.IntFlag.__or__
-        __and__ = enum.IntFlag.__and__
-        __xor__ = enum.IntFlag.__xor__
-        __ror__ = enum.IntFlag.__ror__
-        __rand__ = enum.IntFlag.__rand__
-        __rxor__ = enum.IntFlag.__rxor__
-        __invert__ = enum.IntFlag.__invert__
+        class _NewEnum(int_type, enum.ReprEnum, enum.Flag, boundary=enum.KEEP):
+            pass
+
+    else:
+
+        class _NewEnum(int_type, enum.Flag):
+            # Rebind classmethods to our own class
+            _missing_ = classmethod(enum.IntFlag._missing_.__func__)
+            _create_pseudo_member_ = classmethod(
+                enum.IntFlag._create_pseudo_member_.__func__
+            )
+
+            __or__ = enum.IntFlag.__or__
+            __and__ = enum.IntFlag.__and__
+            __xor__ = enum.IntFlag.__xor__
+            __ror__ = enum.IntFlag.__ror__
+            __rand__ = enum.IntFlag.__rand__
+            __rxor__ = enum.IntFlag.__rxor__
+            __invert__ = enum.IntFlag.__invert__
 
     return _NewEnum
 
@@ -348,6 +464,18 @@ class enum16(enum_factory(uint16_t)):  # noqa: N801
     pass
 
 
+class enum32(enum_factory(uint32_t)):  # noqa: N801
+    pass
+
+
+class enum16_be(enum_factory(uint16_t_be)):  # noqa: N801
+    pass
+
+
+class enum32_be(enum_factory(uint32_t_be)):  # noqa: N801
+    pass
+
+
 class bitmap2(bitmap_factory(uint2_t)):
     pass
 
@@ -401,6 +529,34 @@ class bitmap56(bitmap_factory(uint56_t)):
 
 
 class bitmap64(bitmap_factory(uint64_t)):
+    pass
+
+
+class bitmap16_be(bitmap_factory(uint16_t_be)):
+    pass
+
+
+class bitmap24_be(bitmap_factory(uint24_t_be)):
+    pass
+
+
+class bitmap32_be(bitmap_factory(uint32_t_be)):
+    pass
+
+
+class bitmap40_be(bitmap_factory(uint40_t_be)):
+    pass
+
+
+class bitmap48_be(bitmap_factory(uint48_t_be)):
+    pass
+
+
+class bitmap56_be(bitmap_factory(uint56_t_be)):
+    pass
+
+
+class bitmap64_be(bitmap_factory(uint64_t_be)):
     pass
 
 
