@@ -1,7 +1,7 @@
 """Typing helpers for Zigpy."""
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, Any, Dict
+from typing import TYPE_CHECKING, Any, Dict, Union
 
 ConfigType = Dict[str, Any]
 
@@ -14,12 +14,14 @@ CustomEndpointType = "CustomEndpoint"
 DeviceType = "Device"
 EndpointType = "Endpoint"
 ZDOType = "ZDO"
+AddressingMode = "AddressingMode"
 
 if TYPE_CHECKING:
     import zigpy.application
     import zigpy.device
     import zigpy.endpoint
     import zigpy.quirks
+    import zigpy.types
     import zigpy.zcl
     import zigpy.zdo
 
@@ -31,3 +33,9 @@ if TYPE_CHECKING:
     DeviceType = zigpy.device.Device
     EndpointType = zigpy.endpoint.Endpoint
     ZDOType = zigpy.zdo.ZDO
+
+    AddressingMode = Union[
+        zigpy.types.Addressing.Group,
+        zigpy.types.Addressing.IEEE,
+        zigpy.types.Addressing.NWK,
+    ]
