@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 import zigpy.types as t
-from zigpy.zcl import Cluster
+from zigpy.zcl import Cluster, foundation
 from zigpy.zcl.foundation import ZCLAttributeDef, ZCLCommandDef
 
 
@@ -10,18 +10,30 @@ class ApplianceIdentification(Cluster):
     name = "Appliance Identification"
     ep_attribute = "appliance_id"
     attributes: dict[int, ZCLAttributeDef] = {
-        0x0000: ("basic_identification", t.uint56_t),
-        0x0010: ("company_name", t.LimitedCharString(16)),
-        0x0011: ("company_id", t.uint16_t),
-        0x0012: ("brand_name", t.LimitedCharString(16)),
-        0x0013: ("brand_id", t.uint16_t),
-        0x0014: ("model", t.LimitedLVBytes(16)),
-        0x0015: ("part_number", t.LimitedLVBytes(16)),
-        0x0016: ("product_revision", t.LimitedLVBytes(6)),
-        0x0017: ("software_revision", t.LimitedLVBytes(6)),
-        0x0018: ("product_type_name", t.LVBytesSize2),
-        0x0019: ("product_type_id", t.uint16_t),
-        0x001A: ("ceced_specification_version", t.uint8_t),
+        0x0000: ZCLAttributeDef(
+            "basic_identification", type=t.uint56_t, access="r", mandatory=True
+        ),
+        0x0010: ZCLAttributeDef(
+            "company_name", type=t.LimitedCharString(16), access="r"
+        ),
+        0x0011: ZCLAttributeDef("company_id", type=t.uint16_t, access="r"),
+        0x0012: ZCLAttributeDef("brand_name", type=t.LimitedCharString(16), access="r"),
+        0x0013: ZCLAttributeDef("brand_id", type=t.uint16_t, access="r"),
+        0x0014: ZCLAttributeDef("model", type=t.LimitedLVBytes(16), access="r"),
+        0x0015: ZCLAttributeDef("part_number", type=t.LimitedLVBytes(16), access="r"),
+        0x0016: ZCLAttributeDef(
+            "product_revision", type=t.LimitedLVBytes(6), access="r"
+        ),
+        0x0017: ZCLAttributeDef(
+            "software_revision", type=t.LimitedLVBytes(6), access="r"
+        ),
+        0x0018: ZCLAttributeDef("product_type_name", type=t.LVBytesSize2, access="r"),
+        0x0019: ZCLAttributeDef("product_type_id", type=t.uint16_t, access="r"),
+        0x001A: ZCLAttributeDef(
+            "ceced_specification_version", type=t.uint8_t, access="r"
+        ),
+        0xFFFD: foundation.ZCL_CLUSTER_REVISION_ATTR,
+        0xFFFE: foundation.ZCL_REPORTING_STATUS_ATTR,
     }
     server_commands: dict[int, ZCLCommandDef] = {}
     client_commands: dict[int, ZCLCommandDef] = {}
@@ -32,18 +44,40 @@ class MeterIdentification(Cluster):
     name = "Meter Identification"
     ep_attribute = "meter_id"
     attributes: dict[int, ZCLAttributeDef] = {
-        0x0000: ("company_name", t.LimitedCharString(16)),
-        0x0001: ("meter_type_id", t.uint16_t),
-        0x0004: ("data_quality_id", t.uint16_t),
-        0x0005: ("customer_name", t.LimitedCharString(16)),
-        0x0006: ("model", t.LimitedLVBytes(16)),
-        0x0007: ("part_number", t.LimitedLVBytes(16)),
-        0x0008: ("product_revision", t.LimitedLVBytes(6)),
-        0x000A: ("software_revision", t.LimitedLVBytes(6)),
-        0x000B: ("utility_name", t.LimitedCharString(16)),
-        0x000C: ("pod", t.LimitedCharString(16)),
-        0x000D: ("available_power", t.int24s),
-        0x000E: ("power_threshold", t.int24s),
+        0x0000: ZCLAttributeDef(
+            "company_name", type=t.LimitedCharString(16), access="r", mandatory=True
+        ),
+        0x0001: ZCLAttributeDef(
+            "meter_type_id", type=t.uint16_t, access="r", mandatory=True
+        ),
+        0x0004: ZCLAttributeDef(
+            "data_quality_id", type=t.uint16_t, access="r", mandatory=True
+        ),
+        0x0005: ZCLAttributeDef(
+            "customer_name", type=t.LimitedCharString(16), access="rw"
+        ),
+        0x0006: ZCLAttributeDef("model", type=t.LimitedLVBytes(16), access="r"),
+        0x0007: ZCLAttributeDef("part_number", type=t.LimitedLVBytes(16), access="r"),
+        0x0008: ZCLAttributeDef(
+            "product_revision", type=t.LimitedLVBytes(6), access="r"
+        ),
+        0x000A: ZCLAttributeDef(
+            "software_revision", type=t.LimitedLVBytes(6), access="r"
+        ),
+        0x000B: ZCLAttributeDef(
+            "utility_name", type=t.LimitedCharString(16), access="r"
+        ),
+        0x000C: ZCLAttributeDef(
+            "pod", type=t.LimitedCharString(16), access="r", mandatory=True
+        ),
+        0x000D: ZCLAttributeDef(
+            "available_power", type=t.int24s, access="r", mandatory=True
+        ),
+        0x000E: ZCLAttributeDef(
+            "power_threshold", type=t.int24s, access="r", mandatory=True
+        ),
+        0xFFFD: foundation.ZCL_CLUSTER_REVISION_ATTR,
+        0xFFFE: foundation.ZCL_REPORTING_STATUS_ATTR,
     }
     server_commands: dict[int, ZCLCommandDef] = {}
     client_commands: dict[int, ZCLCommandDef] = {}
@@ -53,7 +87,10 @@ class ApplianceEventAlerts(Cluster):
     cluster_id = 0x0B02
     name = "Appliance Event Alerts"
     ep_attribute = "appliance_event"
-    attributes: dict[int, ZCLAttributeDef] = {}
+    attributes: dict[int, ZCLAttributeDef] = {
+        0xFFFD: foundation.ZCL_CLUSTER_REVISION_ATTR,
+        0xFFFE: foundation.ZCL_REPORTING_STATUS_ATTR,
+    }
     server_commands: dict[int, ZCLCommandDef] = {
         0x00: ZCLCommandDef("get_alerts", {}, False)
     }
@@ -69,8 +106,14 @@ class ApplianceStatistics(Cluster):
     name = "Appliance Statistics"
     ep_attribute = "appliance_stats"
     attributes: dict[int, ZCLAttributeDef] = {
-        0x0000: ("log_max_size", t.uint32_t),
-        0x0001: ("log_queue_max_size", t.uint8_t),
+        0x0000: ZCLAttributeDef(
+            "log_max_size", type=t.uint32_t, access="r", mandatory=True
+        ),
+        0x0001: ZCLAttributeDef(
+            "log_queue_max_size", type=t.uint8_t, access="r", mandatory=True
+        ),
+        0xFFFD: foundation.ZCL_CLUSTER_REVISION_ATTR,
+        0xFFFE: foundation.ZCL_REPORTING_STATUS_ATTR,
     }
     server_commands: dict[int, ZCLCommandDef] = {
         0x00: ZCLCommandDef("log", {}, False),
@@ -88,147 +131,251 @@ class ElectricalMeasurement(Cluster):
     cluster_id = 0x0B04
     name = "Electrical Measurement"
     ep_attribute = "electrical_measurement"
+
+    class MeasurementType(t.bitmap32):
+        Active_measurement_AC = 2 << 0
+        Reactive_measurement_AC = 2 << 1
+        Apparent_measurement_AC = 2 << 2
+        Phase_A_measurement = 2 << 3
+        Phase_B_measurement = 2 << 4
+        Phase_C_measurement = 2 << 5
+        DC_measurement = 2 << 6
+        Harmonics_measurement = 2 << 7
+        Power_quality_measurement = 2 << 8
+
+    class DCOverloadAlarmMark(t.bitmap8):
+        Voltage_Overload = 0b00000001
+        Current_Overload = 0b00000010
+
+    class ACAlarmsMask(t.bitmap16):
+        Voltage_Overload = 2 << 0
+        Current_Overload = 2 << 1
+        Active_Power_Overload = 2 << 2
+        Reactive_Power_Overload = 2 << 3
+        Average_RMS_Over_Voltage = 2 << 4
+        Average_RMS_Under_Voltage = 2 << 5
+        RMS_Extreme_Over_Voltage = 2 << 6
+        RMS_Extreme_Under_Voltage = 2 << 7
+        RMS_Voltage_Sag = 2 << 8
+        RMS_Voltage_Swell = 2 << 9
+
     attributes: dict[int, ZCLAttributeDef] = {
         # Basic Information
-        0x0000: ("measurement_type", t.bitmap32),
+        0x0000: ZCLAttributeDef(
+            "measurement_type", type=MeasurementType, access="r", mandatory=True
+        ),
         # DC Measurement
-        0x0100: ("dc_voltage", t.int16s),
-        0x0101: ("dc_voltage_min", t.int16s),
-        0x0102: ("dc_voltage_max", t.int16s),
-        0x0103: ("dc_current", t.int16s),
-        0x0104: ("dc_current_min", t.int16s),
-        0x0105: ("dc_current_max", t.int16s),
-        0x0106: ("dc_power", t.int16s),
-        0x0107: ("dc_power_min", t.int16s),
-        0x0108: ("dc_power_max", t.int16s),
+        0x0100: ZCLAttributeDef("dc_voltage", type=t.int16s, access="rp"),
+        0x0101: ZCLAttributeDef("dc_voltage_min", type=t.int16s, access="r"),
+        0x0102: ZCLAttributeDef("dc_voltage_max", type=t.int16s, access="r"),
+        0x0103: ZCLAttributeDef("dc_current", type=t.int16s, access="rp"),
+        0x0104: ZCLAttributeDef("dc_current_min", type=t.int16s, access="r"),
+        0x0105: ZCLAttributeDef("dc_current_max", type=t.int16s, access="r"),
+        0x0106: ZCLAttributeDef("dc_power", type=t.int16s, access="rp"),
+        0x0107: ZCLAttributeDef("dc_power_min", type=t.int16s, access="r"),
+        0x0108: ZCLAttributeDef("dc_power_max", type=t.int16s, access="r"),
         # DC Formatting
-        0x0200: ("dc_voltage_multiplier", t.uint16_t),
-        0x0201: ("dc_voltage_divisor", t.uint16_t),
-        0x0202: ("dc_current_multiplier", t.uint16_t),
-        0x0203: ("dc_current_divisor", t.uint16_t),
-        0x0204: ("dc_power_multiplier", t.uint16_t),
-        0x0205: ("dc_power_divisor", t.uint16_t),
+        0x0200: ZCLAttributeDef("dc_voltage_multiplier", type=t.uint16_t, access="rp"),
+        0x0201: ZCLAttributeDef("dc_voltage_divisor", type=t.uint16_t, access="rp"),
+        0x0202: ZCLAttributeDef("dc_current_multiplier", type=t.uint16_t, access="rp"),
+        0x0203: ZCLAttributeDef("dc_current_divisor", type=t.uint16_t, access="rp"),
+        0x0204: ZCLAttributeDef("dc_power_multiplier", type=t.uint16_t, access="rp"),
+        0x0205: ZCLAttributeDef("dc_power_divisor", type=t.uint16_t, access="rp"),
         # AC (Non-phase Specific) Measurements
-        0x0300: ("ac_frequency", t.uint16_t),
-        0x0301: ("ac_frequency_min", t.uint16_t),
-        0x0302: ("ac_frequency_max", t.uint16_t),
-        0x0303: ("neutral_current", t.uint16_t),
-        0x0304: ("total_active_power", t.int32s),
-        0x0305: ("total_reactive_power", t.int32s),
-        0x0306: ("total_apparent_power", t.uint32_t),
-        0x0307: ("meas1st_harmonic_current", t.int16s),
-        0x0308: ("meas3rd_harmonic_current", t.int16s),
-        0x0309: ("meas5th_harmonic_current", t.int16s),
-        0x030A: ("meas7th_harmonic_current", t.int16s),
-        0x030B: ("meas9th_harmonic_current", t.int16s),
-        0x030C: ("meas11th_harmonic_current", t.int16s),
-        0x030D: ("meas_phase1st_harmonic_current", t.int16s),
-        0x030E: ("meas_phase3rd_harmonic_current", t.int16s),
-        0x030F: ("meas_phase5th_harmonic_current", t.int16s),
-        0x0310: ("meas_phase7th_harmonic_current", t.int16s),
-        0x0311: ("meas_phase9th_harmonic_current", t.int16s),
-        0x0312: ("meas_phase11th_harmonic_current", t.int16s),
+        0x0300: ZCLAttributeDef("ac_frequency", type=t.uint16_t, access="rp"),
+        0x0301: ZCLAttributeDef("ac_frequency_min", type=t.uint16_t, access="r"),
+        0x0302: ZCLAttributeDef("ac_frequency_max", type=t.uint16_t, access="r"),
+        0x0303: ZCLAttributeDef("neutral_current", type=t.uint16_t, access="rp"),
+        0x0304: ZCLAttributeDef("total_active_power", type=t.int32s, access="rp"),
+        0x0305: ZCLAttributeDef("total_reactive_power", type=t.int32s, access="rp"),
+        0x0306: ZCLAttributeDef("total_apparent_power", type=t.uint32_t, access="rp"),
+        0x0307: ZCLAttributeDef("meas1st_harmonic_current", type=t.int16s, access="rp"),
+        0x0308: ZCLAttributeDef("meas3rd_harmonic_current", type=t.int16s, access="rp"),
+        0x0309: ZCLAttributeDef("meas5th_harmonic_current", type=t.int16s, access="rp"),
+        0x030A: ZCLAttributeDef("meas7th_harmonic_current", type=t.int16s, access="rp"),
+        0x030B: ZCLAttributeDef("meas9th_harmonic_current", type=t.int16s, access="rp"),
+        0x030C: ZCLAttributeDef(
+            "meas11th_harmonic_current", type=t.int16s, access="rp"
+        ),
+        0x030D: ZCLAttributeDef(
+            "meas_phase1st_harmonic_current", type=t.int16s, access="rp"
+        ),
+        0x030E: ZCLAttributeDef(
+            "meas_phase3rd_harmonic_current", type=t.int16s, access="rp"
+        ),
+        0x030F: ZCLAttributeDef(
+            "meas_phase5th_harmonic_current", type=t.int16s, access="rp"
+        ),
+        0x0310: ZCLAttributeDef(
+            "meas_phase7th_harmonic_current", type=t.int16s, access="rp"
+        ),
+        0x0311: ZCLAttributeDef(
+            "meas_phase9th_harmonic_current", type=t.int16s, access="rp"
+        ),
+        0x0312: ZCLAttributeDef(
+            "meas_phase11th_harmonic_current", type=t.int16s, access="rp"
+        ),
         # AC (Non-phase specific) Formatting
-        0x0400: ("ac_frequency_multiplier", t.uint16_t),
-        0x0401: ("ac_frequency_divisor", t.uint16_t),
-        0x0402: ("power_multiplier", t.uint32_t),
-        0x0403: ("power_divisor", t.uint32_t),
-        0x0404: ("harmonic_current_multiplier", t.int8s),
-        0x0405: ("phase_harmonic_current_multiplier", t.int8s),
+        0x0400: ZCLAttributeDef(
+            "ac_frequency_multiplier", type=t.uint16_t, access="rp"
+        ),
+        0x0401: ZCLAttributeDef("ac_frequency_divisor", type=t.uint16_t, access="rp"),
+        0x0402: ZCLAttributeDef("power_multiplier", type=t.uint32_t, access="rp"),
+        0x0403: ZCLAttributeDef("power_divisor", type=t.uint32_t, access="rp"),
+        0x0404: ZCLAttributeDef(
+            "harmonic_current_multiplier", type=t.int8s, access="rp"
+        ),
+        0x0405: ZCLAttributeDef(
+            "phase_harmonic_current_multiplier", type=t.int8s, access="rp"
+        ),
         # AC (Single Phase or Phase A) Measurements
-        0x0500: ("instantaneous_voltage", t.int16s),
-        0x0501: ("instantaneous_line_current", t.uint16_t),
-        0x0502: ("instantaneous_active_current", t.int16s),
-        0x0503: ("instantaneous_reactive_current", t.int16s),
-        0x0504: ("instantaneous_power", t.int16s),
-        0x0505: ("rms_voltage", t.uint16_t),
-        0x0506: ("rms_voltage_min", t.uint16_t),
-        0x0507: ("rms_voltage_max", t.uint16_t),
-        0x0508: ("rms_current", t.uint16_t),
-        0x0509: ("rms_current_min", t.uint16_t),
-        0x050A: ("rms_current_max", t.uint16_t),
-        0x050B: ("active_power", t.int16s),
-        0x050C: ("active_power_min", t.int16s),
-        0x050D: ("active_power_max", t.int16s),
-        0x050E: ("reactive_power", t.int16s),
-        0x050F: ("apparent_power", t.uint16_t),
-        0x0510: ("power_factor", t.int8s),
-        0x0511: ("average_rms_voltage_meas_period", t.uint16_t),
-        0x0512: ("average_rms_over_voltage_counter", t.uint16_t),
-        0x0513: ("average_rms_under_voltage_counter", t.uint16_t),
-        0x0514: ("rms_extreme_over_voltage_period", t.uint16_t),
-        0x0515: ("rms_extreme_under_voltage_period", t.uint16_t),
-        0x0516: ("rms_voltage_sag_period", t.uint16_t),
-        0x0517: ("rms_voltage_swell_period", t.uint16_t),
+        0x0500: ZCLAttributeDef("instantaneous_voltage", type=t.int16s, access="rp"),
+        0x0501: ZCLAttributeDef(
+            "instantaneous_line_current", type=t.uint16_t, access="rp"
+        ),
+        0x0502: ZCLAttributeDef(
+            "instantaneous_active_current", type=t.int16s, access="rp"
+        ),
+        0x0503: ZCLAttributeDef(
+            "instantaneous_reactive_current", type=t.int16s, access="rp"
+        ),
+        0x0504: ZCLAttributeDef("instantaneous_power", type=t.int16s, access="rp"),
+        0x0505: ZCLAttributeDef("rms_voltage", type=t.uint16_t, access="rp"),
+        0x0506: ZCLAttributeDef("rms_voltage_min", type=t.uint16_t, access="r"),
+        0x0507: ZCLAttributeDef("rms_voltage_max", type=t.uint16_t, access="r"),
+        0x0508: ZCLAttributeDef("rms_current", type=t.uint16_t, access="rp"),
+        0x0509: ZCLAttributeDef("rms_current_min", type=t.uint16_t, access="r"),
+        0x050A: ZCLAttributeDef("rms_current_max", type=t.uint16_t, access="r"),
+        0x050B: ZCLAttributeDef("active_power", type=t.int16s, access="rp"),
+        0x050C: ZCLAttributeDef("active_power_min", type=t.int16s, access="r"),
+        0x050D: ZCLAttributeDef("active_power_max", type=t.int16s, access="r"),
+        0x050E: ZCLAttributeDef("reactive_power", type=t.int16s, access="rp"),
+        0x050F: ZCLAttributeDef("apparent_power", type=t.uint16_t, access="rp"),
+        0x0510: ZCLAttributeDef("power_factor", type=t.int8s, access="r"),
+        0x0511: ZCLAttributeDef(
+            "average_rms_voltage_meas_period", type=t.uint16_t, access="rw"
+        ),
+        0x0512: ZCLAttributeDef(
+            "average_rms_over_voltage_counter", type=t.uint16_t, access="rw"
+        ),
+        0x0513: ZCLAttributeDef(
+            "average_rms_under_voltage_counter", type=t.uint16_t, access="rw"
+        ),
+        0x0514: ZCLAttributeDef(
+            "rms_extreme_over_voltage_period", type=t.uint16_t, access="rw"
+        ),
+        0x0515: ZCLAttributeDef(
+            "rms_extreme_under_voltage_period", type=t.uint16_t, access="rw"
+        ),
+        0x0516: ZCLAttributeDef("rms_voltage_sag_period", type=t.uint16_t, access="rw"),
+        0x0517: ZCLAttributeDef(
+            "rms_voltage_swell_period", type=t.uint16_t, access="rw"
+        ),
         # AC Formatting
-        0x0600: ("ac_voltage_multiplier", t.uint16_t),
-        0x0601: ("ac_voltage_divisor", t.uint16_t),
-        0x0602: ("ac_current_multiplier", t.uint16_t),
-        0x0603: ("ac_current_divisor", t.uint16_t),
-        0x0604: ("ac_power_multiplier", t.uint16_t),
-        0x0605: ("ac_power_divisor", t.uint16_t),
+        0x0600: ZCLAttributeDef("ac_voltage_multiplier", type=t.uint16_t, access="rp"),
+        0x0601: ZCLAttributeDef("ac_voltage_divisor", type=t.uint16_t, access="rp"),
+        0x0602: ZCLAttributeDef("ac_current_multiplier", type=t.uint16_t, access="rp"),
+        0x0603: ZCLAttributeDef("ac_current_divisor", type=t.uint16_t, access="rp"),
+        0x0604: ZCLAttributeDef("ac_power_multiplier", type=t.uint16_t, access="rp"),
+        0x0605: ZCLAttributeDef("ac_power_divisor", type=t.uint16_t, access="rp"),
         # DC Manufacturer Threshold Alarms
-        0x0700: ("dc_overload_alarms_mask", t.bitmap8),
-        0x0701: ("dc_voltage_overload", t.int16s),
-        0x0702: ("dc_current_overload", t.int16s),
+        0x0700: ZCLAttributeDef(
+            "dc_overload_alarms_mask", type=DCOverloadAlarmMark, access="rp"
+        ),
+        0x0701: ZCLAttributeDef("dc_voltage_overload", type=t.int16s, access="rp"),
+        0x0702: ZCLAttributeDef("dc_current_overload", type=t.int16s, access="rp"),
         # AC Manufacturer Threshold Alarms
-        0x0800: ("ac_alarms_mask", t.bitmap16),
-        0x0801: ("ac_voltage_overload", t.int16s),
-        0x0802: ("ac_current_overload", t.int16s),
-        0x0803: ("ac_active_power_overload", t.int16s),
-        0x0804: ("ac_reactive_power_overload", t.int16s),
-        0x0805: ("average_rms_over_voltage", t.int16s),
-        0x0806: ("average_rms_under_voltage", t.int16s),
-        0x0807: ("rms_extreme_over_voltage", t.int16s),
-        0x0808: ("rms_extreme_under_voltage", t.int16s),
-        0x0809: ("rms_voltage_sag", t.int16s),
-        0x080A: ("rms_voltage_swell", t.int16s),
+        0x0800: ZCLAttributeDef("ac_alarms_mask", type=ACAlarmsMask, access="rw"),
+        0x0801: ZCLAttributeDef("ac_voltage_overload", type=t.int16s, access="r"),
+        0x0802: ZCLAttributeDef("ac_current_overload", type=t.int16s, access="r"),
+        0x0803: ZCLAttributeDef("ac_active_power_overload", type=t.int16s, access="r"),
+        0x0804: ZCLAttributeDef(
+            "ac_reactive_power_overload", type=t.int16s, access="r"
+        ),
+        0x0805: ZCLAttributeDef("average_rms_over_voltage", type=t.int16s, access="r"),
+        0x0806: ZCLAttributeDef("average_rms_under_voltage", type=t.int16s, access="r"),
+        0x0807: ZCLAttributeDef("rms_extreme_over_voltage", type=t.int16s, access="rw"),
+        0x0808: ZCLAttributeDef(
+            "rms_extreme_under_voltage", type=t.int16s, access="rw"
+        ),
+        0x0809: ZCLAttributeDef("rms_voltage_sag", type=t.int16s, access="rw"),
+        0x080A: ZCLAttributeDef("rms_voltage_swell", type=t.int16s, access="rw"),
         # AC Phase B Measurements
-        0x0901: ("line_current_ph_b", t.uint16_t),
-        0x0902: ("active_current_ph_b", t.int16s),
-        0x0903: ("reactive_current_ph_b", t.int16s),
-        0x0905: ("rms_voltage_ph_b", t.uint16_t),
-        0x0906: ("rms_voltage_min_ph_b", t.uint16_t),
-        0x0907: ("rms_voltage_max_ph_b", t.uint16_t),
-        0x0908: ("rms_current_ph_b", t.uint16_t),
-        0x0909: ("rms_current_min_ph_b", t.uint16_t),
-        0x090A: ("rms_current_max_ph_b", t.uint16_t),
-        0x090B: ("active_power_ph_b", t.int16s),
-        0x090C: ("active_power_min_ph_b", t.int16s),
-        0x090D: ("active_power_max_ph_b", t.int16s),
-        0x090E: ("reactive_power_ph_b", t.int16s),
-        0x090F: ("apparent_power_ph_b", t.uint16_t),
-        0x0910: ("power_factor_ph_b", t.int8s),
-        0x0911: ("average_rms_voltage_measure_period_ph_b", t.uint16_t),
-        0x0912: ("average_rms_over_voltage_counter_ph_b", t.uint16_t),
-        0x0913: ("average_under_voltage_counter_ph_b", t.uint16_t),
-        0x0914: ("rms_extreme_over_voltage_period_ph_b", t.uint16_t),
-        0x0915: ("rms_extreme_under_voltage_period_ph_b", t.uint16_t),
-        0x0916: ("rms_voltage_sag_period_ph_b", t.uint16_t),
-        0x0917: ("rms_voltage_swell_period_ph_b", t.uint16_t),
+        0x0901: ZCLAttributeDef("line_current_ph_b", type=t.uint16_t, access="rp"),
+        0x0902: ZCLAttributeDef("active_current_ph_b", type=t.int16s, access="rp"),
+        0x0903: ZCLAttributeDef("reactive_current_ph_b", type=t.int16s, access="rp"),
+        0x0905: ZCLAttributeDef("rms_voltage_ph_b", type=t.uint16_t, access="rp"),
+        0x0906: ZCLAttributeDef("rms_voltage_min_ph_b", type=t.uint16_t, access="r"),
+        0x0907: ZCLAttributeDef("rms_voltage_max_ph_b", type=t.uint16_t, access="r"),
+        0x0908: ZCLAttributeDef("rms_current_ph_b", type=t.uint16_t, access="rp"),
+        0x0909: ZCLAttributeDef("rms_current_min_ph_b", type=t.uint16_t, access="r"),
+        0x090A: ZCLAttributeDef("rms_current_max_ph_b", type=t.uint16_t, access="r"),
+        0x090B: ZCLAttributeDef("active_power_ph_b", type=t.int16s, access="rp"),
+        0x090C: ZCLAttributeDef("active_power_min_ph_b", type=t.int16s, access="r"),
+        0x090D: ZCLAttributeDef("active_power_max_ph_b", type=t.int16s, access="r"),
+        0x090E: ZCLAttributeDef("reactive_power_ph_b", type=t.int16s, access="rp"),
+        0x090F: ZCLAttributeDef("apparent_power_ph_b", type=t.uint16_t, access="rp"),
+        0x0910: ZCLAttributeDef("power_factor_ph_b", type=t.int8s, access="r"),
+        0x0911: ZCLAttributeDef(
+            "average_rms_voltage_measure_period_ph_b", type=t.uint16_t, access="rw"
+        ),
+        0x0912: ZCLAttributeDef(
+            "average_rms_over_voltage_counter_ph_b", type=t.uint16_t, access="rw"
+        ),
+        0x0913: ZCLAttributeDef(
+            "average_under_voltage_counter_ph_b", type=t.uint16_t, access="rw"
+        ),
+        0x0914: ZCLAttributeDef(
+            "rms_extreme_over_voltage_period_ph_b", type=t.uint16_t, access="rw"
+        ),
+        0x0915: ZCLAttributeDef(
+            "rms_extreme_under_voltage_period_ph_b", type=t.uint16_t, access="rw"
+        ),
+        0x0916: ZCLAttributeDef(
+            "rms_voltage_sag_period_ph_b", type=t.uint16_t, access="rw"
+        ),
+        0x0917: ZCLAttributeDef(
+            "rms_voltage_swell_period_ph_b", type=t.uint16_t, access="rw"
+        ),
         # AC Phase C Measurements
-        0x0A01: ("line_current_ph_c", t.uint16_t),
-        0x0A02: ("active_current_ph_c", t.int16s),
-        0x0A03: ("reactive_current_ph_c", t.int16s),
-        0x0A05: ("rms_voltage_ph_c", t.uint16_t),
-        0x0A06: ("rms_voltage_min_ph_c", t.uint16_t),
-        0x0A07: ("rms_voltage_max_ph_c", t.uint16_t),
-        0x0A08: ("rms_current_ph_c", t.uint16_t),
-        0x0A09: ("rms_current_min_ph_c", t.uint16_t),
-        0x0A0A: ("rms_current_max_ph_c", t.uint16_t),
-        0x0A0B: ("active_power_ph_c", t.int16s),
-        0x0A0C: ("active_power_min_ph_c", t.int16s),
-        0x0A0D: ("active_power_max_ph_c", t.int16s),
-        0x0A0E: ("reactive_power_ph_c", t.int16s),
-        0x0A0F: ("apparent_power_ph_c", t.uint16_t),
-        0x0A10: ("power_factor_ph_c", t.int8s),
-        0x0A11: ("average_rms_voltage_meas_period_ph_c", t.uint16_t),
-        0x0A12: ("average_rms_over_voltage_counter_ph_c", t.uint16_t),
-        0x0A13: ("average_under_voltage_counter_ph_c", t.uint16_t),
-        0x0A14: ("rms_extreme_over_voltage_period_ph_c", t.uint16_t),
-        0x0A15: ("rms_extreme_under_voltage_period_ph_c", t.uint16_t),
-        0x0A16: ("rms_voltage_sag_period_ph_c", t.uint16_t),
-        0x0A17: ("rms_voltage_swell_period_ph__c", t.uint16_t),
+        0x0A01: ZCLAttributeDef("line_current_ph_c", type=t.uint16_t, access="rp"),
+        0x0A02: ZCLAttributeDef("active_current_ph_c", type=t.int16s, access="rp"),
+        0x0A03: ZCLAttributeDef("reactive_current_ph_c", type=t.int16s, access="rp"),
+        0x0A05: ZCLAttributeDef("rms_voltage_ph_c", type=t.uint16_t, access="rp"),
+        0x0A06: ZCLAttributeDef("rms_voltage_min_ph_c", type=t.uint16_t, access="r"),
+        0x0A07: ZCLAttributeDef("rms_voltage_max_ph_c", type=t.uint16_t, access="r"),
+        0x0A08: ZCLAttributeDef("rms_current_ph_c", type=t.uint16_t, access="rp"),
+        0x0A09: ZCLAttributeDef("rms_current_min_ph_c", type=t.uint16_t, access="r"),
+        0x0A0A: ZCLAttributeDef("rms_current_max_ph_c", type=t.uint16_t, access="r"),
+        0x0A0B: ZCLAttributeDef("active_power_ph_c", type=t.int16s, access="rp"),
+        0x0A0C: ZCLAttributeDef("active_power_min_ph_c", type=t.int16s, access="r"),
+        0x0A0D: ZCLAttributeDef("active_power_max_ph_c", type=t.int16s, access="r"),
+        0x0A0E: ZCLAttributeDef("reactive_power_ph_c", type=t.int16s, access="rp"),
+        0x0A0F: ZCLAttributeDef("apparent_power_ph_c", type=t.uint16_t, access="rp"),
+        0x0A10: ZCLAttributeDef("power_factor_ph_c", type=t.int8s, access="r"),
+        0x0A11: ZCLAttributeDef(
+            "average_rms_voltage_meas_period_ph_c", type=t.uint16_t, access="rw"
+        ),
+        0x0A12: ZCLAttributeDef(
+            "average_rms_over_voltage_counter_ph_c", type=t.uint16_t, access="rw"
+        ),
+        0x0A13: ZCLAttributeDef(
+            "average_under_voltage_counter_ph_c", type=t.uint16_t, access="rw"
+        ),
+        0x0A14: ZCLAttributeDef(
+            "rms_extreme_over_voltage_period_ph_c", type=t.uint16_t, access="rw"
+        ),
+        0x0A15: ZCLAttributeDef(
+            "rms_extreme_under_voltage_period_ph_c", type=t.uint16_t, access="rw"
+        ),
+        0x0A16: ZCLAttributeDef(
+            "rms_voltage_sag_period_ph_c", type=t.uint16_t, access="rw"
+        ),
+        0x0A17: ZCLAttributeDef(
+            "rms_voltage_swell_period_ph_c", type=t.uint16_t, access="rw"
+        ),
+        0xFFFD: foundation.ZCL_CLUSTER_REVISION_ATTR,
+        0xFFFE: foundation.ZCL_REPORTING_STATUS_ATTR,
     }
     server_commands: dict[int, ZCLCommandDef] = {
         0x00: ZCLCommandDef("get_profile_info", {}, False),
@@ -245,39 +392,51 @@ class Diagnostic(Cluster):
     ep_attribute = "diagnostic"
     attributes: dict[int, ZCLAttributeDef] = {
         # Hardware Information
-        0x0000: ("number_of_resets", t.uint16_t),
-        0x0001: ("persistent_memory_writes", t.uint16_t),
+        0x0000: ZCLAttributeDef("number_of_resets", type=t.uint16_t, access="r"),
+        0x0001: ZCLAttributeDef(
+            "persistent_memory_writes", type=t.uint16_t, access="r"
+        ),
         # Stack/Network Information
-        0x0100: ("mac_rx_bcast", t.uint32_t),
-        0x0101: ("mac_tx_bcast", t.uint32_t),
-        0x0102: ("mac_rx_ucast", t.uint32_t),
-        0x0103: ("mac_tx_ucast", t.uint32_t),
-        0x0104: ("mac_tx_ucast_retry", t.uint16_t),
-        0x0105: ("mac_tx_ucast_fail", t.uint16_t),
-        0x0106: ("aps_rx_bcast", t.uint16_t),
-        0x0107: ("aps_tx_bcast", t.uint16_t),
-        0x0108: ("aps_rx_ucast", t.uint16_t),
-        0x0109: ("aps_tx_ucast_success", t.uint16_t),
-        0x010A: ("aps_tx_ucast_retry", t.uint16_t),
-        0x010B: ("aps_tx_ucast_fail", t.uint16_t),
-        0x010C: ("route_disc_initiated", t.uint16_t),
-        0x010D: ("neighbor_added", t.uint16_t),
-        0x010E: ("neighbor_removed", t.uint16_t),
-        0x010F: ("neighbor_stale", t.uint16_t),
-        0x0110: ("join_indication", t.uint16_t),
-        0x0111: ("child_moved", t.uint16_t),
-        0x0112: ("nwk_fc_failure", t.uint16_t),
-        0x0113: ("aps_fc_failure", t.uint16_t),
-        0x0114: ("aps_unauthorized_key", t.uint16_t),
-        0x0115: ("nwk_decrypt_failures", t.uint16_t),
-        0x0116: ("aps_decrypt_failures", t.uint16_t),
-        0x0117: ("packet_buffer_allocate_failures", t.uint16_t),
-        0x0118: ("relayed_ucast", t.uint16_t),
-        0x0119: ("phy_to_mac_queue_limit_reached", t.uint16_t),
-        0x011A: ("packet_validate_drop_count", t.uint16_t),
-        0x011B: ("average_mac_retry_per_aps_message_sent", t.uint16_t),
-        0x011C: ("last_message_lqi", t.uint8_t),
-        0x011D: ("last_message_rssi", t.int8s),
+        0x0100: ZCLAttributeDef("mac_rx_bcast", type=t.uint32_t, access="r"),
+        0x0101: ZCLAttributeDef("mac_tx_bcast", type=t.uint32_t, access="r"),
+        0x0102: ZCLAttributeDef("mac_rx_ucast", type=t.uint32_t, access="r"),
+        0x0103: ZCLAttributeDef("mac_tx_ucast", type=t.uint32_t, access="r"),
+        0x0104: ZCLAttributeDef("mac_tx_ucast_retry", type=t.uint16_t, access="r"),
+        0x0105: ZCLAttributeDef("mac_tx_ucast_fail", type=t.uint16_t, access="r"),
+        0x0106: ZCLAttributeDef("aps_rx_bcast", type=t.uint16_t, access="r"),
+        0x0107: ZCLAttributeDef("aps_tx_bcast", type=t.uint16_t, access="r"),
+        0x0108: ZCLAttributeDef("aps_rx_ucast", type=t.uint16_t, access="r"),
+        0x0109: ZCLAttributeDef("aps_tx_ucast_success", type=t.uint16_t, access="r"),
+        0x010A: ZCLAttributeDef("aps_tx_ucast_retry", type=t.uint16_t, access="r"),
+        0x010B: ZCLAttributeDef("aps_tx_ucast_fail", type=t.uint16_t, access="r"),
+        0x010C: ZCLAttributeDef("route_disc_initiated", type=t.uint16_t, access="r"),
+        0x010D: ZCLAttributeDef("neighbor_added", type=t.uint16_t, access="r"),
+        0x010E: ZCLAttributeDef("neighbor_removed", type=t.uint16_t, access="r"),
+        0x010F: ZCLAttributeDef("neighbor_stale", type=t.uint16_t, access="r"),
+        0x0110: ZCLAttributeDef("join_indication", type=t.uint16_t, access="r"),
+        0x0111: ZCLAttributeDef("child_moved", type=t.uint16_t, access="r"),
+        0x0112: ZCLAttributeDef("nwk_fc_failure", type=t.uint16_t, access="r"),
+        0x0113: ZCLAttributeDef("aps_fc_failure", type=t.uint16_t, access="r"),
+        0x0114: ZCLAttributeDef("aps_unauthorized_key", type=t.uint16_t, access="r"),
+        0x0115: ZCLAttributeDef("nwk_decrypt_failures", type=t.uint16_t, access="r"),
+        0x0116: ZCLAttributeDef("aps_decrypt_failures", type=t.uint16_t, access="r"),
+        0x0117: ZCLAttributeDef(
+            "packet_buffer_allocate_failures", type=t.uint16_t, access="r"
+        ),
+        0x0118: ZCLAttributeDef("relayed_ucast", type=t.uint16_t, access="r"),
+        0x0119: ZCLAttributeDef(
+            "phy_to_mac_queue_limit_reached", type=t.uint16_t, access="r"
+        ),
+        0x011A: ZCLAttributeDef(
+            "packet_validate_drop_count", type=t.uint16_t, access="r"
+        ),
+        0x011B: ZCLAttributeDef(
+            "average_mac_retry_per_aps_message_sent", type=t.uint16_t, access="r"
+        ),
+        0x011C: ZCLAttributeDef("last_message_lqi", type=t.uint8_t, access="r"),
+        0x011D: ZCLAttributeDef("last_message_rssi", type=t.int8s, access="r"),
+        0xFFFD: foundation.ZCL_CLUSTER_REVISION_ATTR,
+        0xFFFE: foundation.ZCL_REPORTING_STATUS_ATTR,
     }
     server_commands: dict[int, ZCLCommandDef] = {}
     client_commands: dict[int, ZCLCommandDef] = {}
