@@ -16,9 +16,9 @@ class GenericTunnel(Cluster):
     cluster_id = 0x0600
     ep_attribute = "generic_tunnel"
     attributes: dict[int, ZCLAttributeDef] = {
-        0x0001: ("max_income_trans_size", t.uint16_t),
-        0x0002: ("max_outgo_trans_size", t.uint16_t),
-        0x0003: ("protocol_addr", t.LVBytes),
+        0x0001: ZCLAttributeDef("max_income_trans_size", type=t.uint16_t),
+        0x0002: ZCLAttributeDef("max_outgo_trans_size", type=t.uint16_t),
+        0x0003: ZCLAttributeDef("protocol_addr", type=t.LVBytes),
     }
     server_commands: dict[int, ZCLCommandDef] = {
         0x00: ZCLCommandDef("match_protocol_addr", {}, False)
@@ -43,13 +43,13 @@ class AnalogInputRegular(Cluster):
     cluster_id = 0x0602
     ep_attribute = "bacnet_regular_analog_input"
     attributes: dict[int, ZCLAttributeDef] = {
-        0x0016: ("cov_increment", t.Single),
-        0x001F: ("device_type", t.CharacterString),
-        0x004B: ("object_id", t.FixedList[4, t.uint8_t]),
-        0x004D: ("object_name", t.CharacterString),
-        0x004F: ("object_type", t.enum16),
-        0x0076: ("update_interval", t.uint8_t),
-        0x00A8: ("profile_name", t.CharacterString),
+        0x0016: ZCLAttributeDef("cov_increment", type=t.Single),
+        0x001F: ZCLAttributeDef("device_type", type=t.CharacterString),
+        0x004B: ZCLAttributeDef("object_id", type=t.FixedList[4, t.uint8_t]),
+        0x004D: ZCLAttributeDef("object_name", type=t.CharacterString),
+        0x004F: ZCLAttributeDef("object_type", type=t.enum16),
+        0x0076: ZCLAttributeDef("update_interval", type=t.uint8_t),
+        0x00A8: ZCLAttributeDef("profile_name", type=t.CharacterString),
     }
     server_commands: dict[int, ZCLCommandDef] = {}
     client_commands: dict[int, ZCLCommandDef] = {}
@@ -59,17 +59,17 @@ class AnalogInputExtended(Cluster):
     cluster_id = 0x0603
     ep_attribute = "bacnet_extended_analog_input"
     attributes: dict[int, ZCLAttributeDef] = {
-        0x0000: ("acked_transitions", t.bitmap8),
-        0x0011: ("notification_class", t.uint16_t),
-        0x0019: ("deadband", t.Single),
-        0x0023: ("event_enable", t.bitmap8),
-        0x0024: ("event_state", t.enum8),
-        0x002D: ("high_limit", t.Single),
-        0x0034: ("limit_enable", t.bitmap8),
-        0x003B: ("low_limit", t.Single),
-        0x0048: ("notify_type", t.enum8),
-        0x0071: ("time_delay", t.uint8_t),
-        # 0x0082: ('event_time_stamps', TODO.array) # Array[3] of (16-bit unsigned
+        0x0000: ZCLAttributeDef("acked_transitions", type=t.bitmap8),
+        0x0011: ZCLAttributeDef("notification_class", type=t.uint16_t),
+        0x0019: ZCLAttributeDef("deadband", type=t.Single),
+        0x0023: ZCLAttributeDef("event_enable", type=t.bitmap8),
+        0x0024: ZCLAttributeDef("event_state", type=t.enum8),
+        0x002D: ZCLAttributeDef("high_limit", type=t.Single),
+        0x0034: ZCLAttributeDef("limit_enable", type=t.bitmap8),
+        0x003B: ZCLAttributeDef("low_limit", type=t.Single),
+        0x0048: ZCLAttributeDef("notify_type", type=t.enum8),
+        0x0071: ZCLAttributeDef("time_delay", type=t.uint8_t),
+        # 0x0082: ZCLAttributeDef('event_time_stamps', type=TODO.array) # Array[3] of (16-bit unsigned
         # integer, time of day, or structure of (date, time of day))
     }
     server_commands: dict[int, ZCLCommandDef] = {
@@ -85,13 +85,13 @@ class AnalogOutputRegular(Cluster):
     cluster_id = 0x0604
     ep_attribute = "bacnet_regular_analog_output"
     attributes: dict[int, ZCLAttributeDef] = {
-        0x0016: ("cov_increment", t.Single),
-        0x001F: ("device_type", t.CharacterString),
-        0x004B: ("object_id", t.FixedList[4, t.uint8_t]),
-        0x004D: ("object_name", t.CharacterString),
-        0x004F: ("object_type", t.enum16),
-        0x0076: ("update_interval", t.uint8_t),
-        0x00A8: ("profile_name", t.CharacterString),
+        0x0016: ZCLAttributeDef("cov_increment", type=t.Single),
+        0x001F: ZCLAttributeDef("device_type", type=t.CharacterString),
+        0x004B: ZCLAttributeDef("object_id", type=t.FixedList[4, t.uint8_t]),
+        0x004D: ZCLAttributeDef("object_name", type=t.CharacterString),
+        0x004F: ZCLAttributeDef("object_type", type=t.enum16),
+        0x0076: ZCLAttributeDef("update_interval", type=t.uint8_t),
+        0x00A8: ZCLAttributeDef("profile_name", type=t.CharacterString),
     }
     server_commands: dict[int, ZCLCommandDef] = {}
     client_commands: dict[int, ZCLCommandDef] = {}
@@ -101,17 +101,17 @@ class AnalogOutputExtended(Cluster):
     cluster_id = 0x0605
     ep_attribute = "bacnet_extended_analog_output"
     attributes: dict[int, ZCLAttributeDef] = {
-        0x0000: ("acked_transitions", t.bitmap8),
-        0x0011: ("notification_class", t.uint16_t),
-        0x0019: ("deadband", t.Single),
-        0x0023: ("event_enable", t.bitmap8),
-        0x0024: ("event_state", t.enum8),
-        0x002D: ("high_limit", t.Single),
-        0x0034: ("limit_enable", t.bitmap8),
-        0x003B: ("low_limit", t.Single),
-        0x0048: ("notify_type", t.enum8),
-        0x0071: ("time_delay", t.uint8_t),
-        # 0x0082: ('event_time_stamps', TODO.array)# Array[3] of (16-bit unsigned
+        0x0000: ZCLAttributeDef("acked_transitions", type=t.bitmap8),
+        0x0011: ZCLAttributeDef("notification_class", type=t.uint16_t),
+        0x0019: ZCLAttributeDef("deadband", type=t.Single),
+        0x0023: ZCLAttributeDef("event_enable", type=t.bitmap8),
+        0x0024: ZCLAttributeDef("event_state", type=t.enum8),
+        0x002D: ZCLAttributeDef("high_limit", type=t.Single),
+        0x0034: ZCLAttributeDef("limit_enable", type=t.bitmap8),
+        0x003B: ZCLAttributeDef("low_limit", type=t.Single),
+        0x0048: ZCLAttributeDef("notify_type", type=t.enum8),
+        0x0071: ZCLAttributeDef("time_delay", type=t.uint8_t),
+        # 0x0082: ZCLAttributeDef('event_time_stamps', type=TODO.array)# Array[3] of (16-bit unsigned
         # integer, time of day, or structure of (date, time of day))
     }
     server_commands: dict[int, ZCLCommandDef] = {}
@@ -122,11 +122,11 @@ class AnalogValueRegular(Cluster):
     cluster_id = 0x0606
     ep_attribute = "bacnet_regular_analog_value"
     attributes: dict[int, ZCLAttributeDef] = {
-        0x0016: ("cov_increment", t.Single),
-        0x004B: ("object_id", t.FixedList[4, t.uint8_t]),
-        0x004D: ("object_name", t.CharacterString),
-        0x004F: ("object_type", t.enum16),
-        0x00A8: ("profile_name", t.CharacterString),
+        0x0016: ZCLAttributeDef("cov_increment", type=t.Single),
+        0x004B: ZCLAttributeDef("object_id", type=t.FixedList[4, t.uint8_t]),
+        0x004D: ZCLAttributeDef("object_name", type=t.CharacterString),
+        0x004F: ZCLAttributeDef("object_type", type=t.enum16),
+        0x00A8: ZCLAttributeDef("profile_name", type=t.CharacterString),
     }
     server_commands: dict[int, ZCLCommandDef] = {}
     client_commands: dict[int, ZCLCommandDef] = {}
@@ -136,17 +136,17 @@ class AnalogValueExtended(Cluster):
     cluster_id = 0x0607
     ep_attribute = "bacnet_extended_analog_value"
     attributes: dict[int, ZCLAttributeDef] = {
-        0x0000: ("acked_transitions", t.bitmap8),
-        0x0011: ("notification_class", t.uint16_t),
-        0x0019: ("deadband", t.Single),
-        0x0023: ("event_enable", t.bitmap8),
-        0x0024: ("event_state", t.enum8),
-        0x002D: ("high_limit", t.Single),
-        0x0034: ("limit_enable", t.bitmap8),
-        0x003B: ("low_limit", t.Single),
-        0x0048: ("notify_type", t.enum8),
-        0x0071: ("time_delay", t.uint8_t),
-        # 0x0082: ('event_time_stamps', TODO.array),  # Array[3] of (16-bit unsigned
+        0x0000: ZCLAttributeDef("acked_transitions", type=t.bitmap8),
+        0x0011: ZCLAttributeDef("notification_class", type=t.uint16_t),
+        0x0019: ZCLAttributeDef("deadband", type=t.Single),
+        0x0023: ZCLAttributeDef("event_enable", type=t.bitmap8),
+        0x0024: ZCLAttributeDef("event_state", type=t.enum8),
+        0x002D: ZCLAttributeDef("high_limit", type=t.Single),
+        0x0034: ZCLAttributeDef("limit_enable", type=t.bitmap8),
+        0x003B: ZCLAttributeDef("low_limit", type=t.Single),
+        0x0048: ZCLAttributeDef("notify_type", type=t.enum8),
+        0x0071: ZCLAttributeDef("time_delay", type=t.uint8_t),
+        # 0x0082: ZCLAttributeDef('event_time_stamps', type=TODO.array),  # Array[3] of (16-bit unsigned
         # integer, time of day, or structure of (date, time of day))
     }
     server_commands: dict[int, ZCLCommandDef] = {}
@@ -157,16 +157,16 @@ class BinaryInputRegular(Cluster):
     cluster_id = 0x0608
     ep_attribute = "bacnet_regular_binary_input"
     attributes: dict[int, ZCLAttributeDef] = {
-        0x000F: ("change_of_state_count", t.uint32_t),
-        0x0010: ("change_of_state_time", DateTime),
-        0x001F: ("device_type", t.CharacterString),
-        0x0021: ("elapsed_active_time", t.uint32_t),
-        0x004B: ("object_id", t.FixedList[4, t.uint8_t]),
-        0x004D: ("object_name", t.CharacterString),
-        0x004F: ("object_type", t.enum16),
-        0x0072: ("time_of_at_reset", DateTime),
-        0x0073: ("time_of_sc_reset", DateTime),
-        0x00A8: ("profile_name", t.CharacterString),
+        0x000F: ZCLAttributeDef("change_of_state_count", type=t.uint32_t),
+        0x0010: ZCLAttributeDef("change_of_state_time", type=DateTime),
+        0x001F: ZCLAttributeDef("device_type", type=t.CharacterString),
+        0x0021: ZCLAttributeDef("elapsed_active_time", type=t.uint32_t),
+        0x004B: ZCLAttributeDef("object_id", type=t.FixedList[4, t.uint8_t]),
+        0x004D: ZCLAttributeDef("object_name", type=t.CharacterString),
+        0x004F: ZCLAttributeDef("object_type", type=t.enum16),
+        0x0072: ZCLAttributeDef("time_of_at_reset", type=DateTime),
+        0x0073: ZCLAttributeDef("time_of_sc_reset", type=DateTime),
+        0x00A8: ZCLAttributeDef("profile_name", type=t.CharacterString),
     }
     server_commands: dict[int, ZCLCommandDef] = {}
     client_commands: dict[int, ZCLCommandDef] = {}
@@ -176,14 +176,14 @@ class BinaryInputExtended(Cluster):
     cluster_id = 0x0609
     ep_attribute = "bacnet_extended_binary_input"
     attributes: dict[int, ZCLAttributeDef] = {
-        0x0000: ("acked_transitions", t.bitmap8),
-        0x0006: ("alarm_value", t.Bool),
-        0x0011: ("notification_class", t.uint16_t),
-        0x0023: ("event_enable", t.bitmap8),
-        0x0024: ("event_state", t.enum8),
-        0x0048: ("notify_type", t.enum8),
-        0x0071: ("time_delay", t.uint8_t),
-        # 0x0082: ('event_time_stamps', TODO.array),  # Array[3] of (16-bit unsigned
+        0x0000: ZCLAttributeDef("acked_transitions", type=t.bitmap8),
+        0x0006: ZCLAttributeDef("alarm_value", type=t.Bool),
+        0x0011: ZCLAttributeDef("notification_class", type=t.uint16_t),
+        0x0023: ZCLAttributeDef("event_enable", type=t.bitmap8),
+        0x0024: ZCLAttributeDef("event_state", type=t.enum8),
+        0x0048: ZCLAttributeDef("notify_type", type=t.enum8),
+        0x0071: ZCLAttributeDef("time_delay", type=t.uint8_t),
+        # 0x0082: ZCLAttributeDef('event_time_stamps', type=TODO.array),  # Array[3] of (16-bit unsigned
         # integer, time of day, or structure of (date, time of day))
     }
     server_commands: dict[int, ZCLCommandDef] = {}
@@ -194,17 +194,17 @@ class BinaryOutputRegular(Cluster):
     cluster_id = 0x060A
     ep_attribute = "bacnet_regular_binary_output"
     attributes: dict[int, ZCLAttributeDef] = {
-        0x000F: ("change_of_state_count", t.uint32_t),
-        0x0010: ("change_of_state_time", DateTime),
-        0x001F: ("device_type", t.CharacterString),
-        0x0021: ("elapsed_active_time", t.uint32_t),
-        0x0028: ("feed_back_value", t.enum8),
-        0x004B: ("object_id", t.FixedList[4, t.uint8_t]),
-        0x004D: ("object_name", t.CharacterString),
-        0x004F: ("object_type", t.enum16),
-        0x0072: ("time_of_at_reset", DateTime),
-        0x0073: ("time_of_sc_reset", DateTime),
-        0x00A8: ("profile_name", t.CharacterString),
+        0x000F: ZCLAttributeDef("change_of_state_count", type=t.uint32_t),
+        0x0010: ZCLAttributeDef("change_of_state_time", type=DateTime),
+        0x001F: ZCLAttributeDef("device_type", type=t.CharacterString),
+        0x0021: ZCLAttributeDef("elapsed_active_time", type=t.uint32_t),
+        0x0028: ZCLAttributeDef("feed_back_value", type=t.enum8),
+        0x004B: ZCLAttributeDef("object_id", type=t.FixedList[4, t.uint8_t]),
+        0x004D: ZCLAttributeDef("object_name", type=t.CharacterString),
+        0x004F: ZCLAttributeDef("object_type", type=t.enum16),
+        0x0072: ZCLAttributeDef("time_of_at_reset", type=DateTime),
+        0x0073: ZCLAttributeDef("time_of_sc_reset", type=DateTime),
+        0x00A8: ZCLAttributeDef("profile_name", type=t.CharacterString),
     }
     server_commands: dict[int, ZCLCommandDef] = {}
     client_commands: dict[int, ZCLCommandDef] = {}
@@ -214,13 +214,13 @@ class BinaryOutputExtended(Cluster):
     cluster_id = 0x060B
     ep_attribute = "bacnet_extended_binary_output"
     attributes: dict[int, ZCLAttributeDef] = {
-        0x0000: ("acked_transitions", t.bitmap8),
-        0x0011: ("notification_class", t.uint16_t),
-        0x0023: ("event_enable", t.bitmap8),
-        0x0024: ("event_state", t.enum8),
-        0x0048: ("notify_type", t.enum8),
-        0x0071: ("time_delay", t.uint8_t),
-        # 0x0082: ('event_time_stamps', TODO.array),  # Array[3] of (16-bit unsigned
+        0x0000: ZCLAttributeDef("acked_transitions", type=t.bitmap8),
+        0x0011: ZCLAttributeDef("notification_class", type=t.uint16_t),
+        0x0023: ZCLAttributeDef("event_enable", type=t.bitmap8),
+        0x0024: ZCLAttributeDef("event_state", type=t.enum8),
+        0x0048: ZCLAttributeDef("notify_type", type=t.enum8),
+        0x0071: ZCLAttributeDef("time_delay", type=t.uint8_t),
+        # 0x0082: ZCLAttributeDef('event_time_stamps', type=TODO.array),  # Array[3] of (16-bit unsigned
         # integer, time of day, or structure of (date, time of day))
     }
     server_commands: dict[int, ZCLCommandDef] = {}
@@ -231,15 +231,15 @@ class BinaryValueRegular(Cluster):
     cluster_id = 0x060C
     ep_attribute = "bacnet_regular_binary_value"
     attributes: dict[int, ZCLAttributeDef] = {
-        0x000F: ("change_of_state_count", t.uint32_t),
-        0x0010: ("change_of_state_time", DateTime),
-        0x0021: ("elapsed_active_time", t.uint32_t),
-        0x004B: ("object_id", t.FixedList[4, t.uint8_t]),
-        0x004D: ("object_name", t.CharacterString),
-        0x004F: ("object_type", t.enum16),
-        0x0072: ("time_of_at_reset", DateTime),
-        0x0073: ("time_of_sc_reset", DateTime),
-        0x00A8: ("profile_name", t.CharacterString),
+        0x000F: ZCLAttributeDef("change_of_state_count", type=t.uint32_t),
+        0x0010: ZCLAttributeDef("change_of_state_time", type=DateTime),
+        0x0021: ZCLAttributeDef("elapsed_active_time", type=t.uint32_t),
+        0x004B: ZCLAttributeDef("object_id", type=t.FixedList[4, t.uint8_t]),
+        0x004D: ZCLAttributeDef("object_name", type=t.CharacterString),
+        0x004F: ZCLAttributeDef("object_type", type=t.enum16),
+        0x0072: ZCLAttributeDef("time_of_at_reset", type=DateTime),
+        0x0073: ZCLAttributeDef("time_of_sc_reset", type=DateTime),
+        0x00A8: ZCLAttributeDef("profile_name", type=t.CharacterString),
     }
     server_commands: dict[int, ZCLCommandDef] = {}
     client_commands: dict[int, ZCLCommandDef] = {}
@@ -249,14 +249,14 @@ class BinaryValueExtended(Cluster):
     cluster_id = 0x060D
     ep_attribute = "bacnet_extended_binary_value"
     attributes: dict[int, ZCLAttributeDef] = {
-        0x0000: ("acked_transitions", t.bitmap8),
-        0x0006: ("alarm_value", t.Bool),
-        0x0011: ("notification_class", t.uint16_t),
-        0x0023: ("event_enable", t.bitmap8),
-        0x0024: ("event_state", t.enum8),
-        0x0048: ("notify_type", t.enum8),
-        0x0071: ("time_delay", t.uint8_t),
-        # 0x0082: ('event_time_stamps', TODO.array),  # Array[3] of (16-bit unsigned
+        0x0000: ZCLAttributeDef("acked_transitions", type=t.bitmap8),
+        0x0006: ZCLAttributeDef("alarm_value", type=t.Bool),
+        0x0011: ZCLAttributeDef("notification_class", type=t.uint16_t),
+        0x0023: ZCLAttributeDef("event_enable", type=t.bitmap8),
+        0x0024: ZCLAttributeDef("event_state", type=t.enum8),
+        0x0048: ZCLAttributeDef("notify_type", type=t.enum8),
+        0x0071: ZCLAttributeDef("time_delay", type=t.uint8_t),
+        # 0x0082: ZCLAttributeDef('event_time_stamps', type=TODO.array),  # Array[3] of (16-bit unsigned
         # integer, time of day, or structure of (date, time of day))
     }
     server_commands: dict[int, ZCLCommandDef] = {}
@@ -267,11 +267,11 @@ class MultistateInputRegular(Cluster):
     cluster_id = 0x060E
     ep_attribute = "bacnet_regular_multistate_input"
     attributes: dict[int, ZCLAttributeDef] = {
-        0x001F: ("device_type", t.CharacterString),
-        0x004B: ("object_id", t.FixedList[4, t.uint8_t]),
-        0x004D: ("object_name", t.CharacterString),
-        0x004F: ("object_type", t.enum16),
-        0x00A8: ("profile_name", t.CharacterString),
+        0x001F: ZCLAttributeDef("device_type", type=t.CharacterString),
+        0x004B: ZCLAttributeDef("object_id", type=t.FixedList[4, t.uint8_t]),
+        0x004D: ZCLAttributeDef("object_name", type=t.CharacterString),
+        0x004F: ZCLAttributeDef("object_type", type=t.enum16),
+        0x00A8: ZCLAttributeDef("profile_name", type=t.CharacterString),
     }
     server_commands: dict[int, ZCLCommandDef] = {}
     client_commands: dict[int, ZCLCommandDef] = {}
@@ -281,15 +281,15 @@ class MultistateInputExtended(Cluster):
     cluster_id = 0x060F
     ep_attribute = "bacnet_extended_multistate_input"
     attributes: dict[int, ZCLAttributeDef] = {
-        0x0000: ("acked_transitions", t.bitmap8),
-        0x0006: ("alarm_value", t.uint16_t),
-        0x0011: ("notification_class", t.uint16_t),
-        0x0023: ("event_enable", t.bitmap8),
-        0x0024: ("event_state", t.enum8),
-        0x0025: ("fault_values", t.uint16_t),
-        0x0048: ("notify_type", t.enum8),
-        0x0071: ("time_delay", t.uint8_t),
-        # 0x0082: ('event_time_stamps', TODO.array),  # Array[3] of (16-bit unsigned
+        0x0000: ZCLAttributeDef("acked_transitions", type=t.bitmap8),
+        0x0006: ZCLAttributeDef("alarm_value", type=t.uint16_t),
+        0x0011: ZCLAttributeDef("notification_class", type=t.uint16_t),
+        0x0023: ZCLAttributeDef("event_enable", type=t.bitmap8),
+        0x0024: ZCLAttributeDef("event_state", type=t.enum8),
+        0x0025: ZCLAttributeDef("fault_values", type=t.uint16_t),
+        0x0048: ZCLAttributeDef("notify_type", type=t.enum8),
+        0x0071: ZCLAttributeDef("time_delay", type=t.uint8_t),
+        # 0x0082: ZCLAttributeDef('event_time_stamps', type=TODO.array),  # Array[3] of (16-bit unsigned
         # integer, time of day, or structure of (date, time of day))
     }
     server_commands: dict[int, ZCLCommandDef] = {}
@@ -300,12 +300,12 @@ class MultistateOutputRegular(Cluster):
     cluster_id = 0x0610
     ep_attribute = "bacnet_regular_multistate_output"
     attributes: dict[int, ZCLAttributeDef] = {
-        0x001F: ("device_type", t.CharacterString),
-        0x0028: ("feed_back_value", t.enum8),
-        0x004B: ("object_id", t.FixedList[4, t.uint8_t]),
-        0x004D: ("object_name", t.CharacterString),
-        0x004F: ("object_type", t.enum16),
-        0x00A8: ("profile_name", t.CharacterString),
+        0x001F: ZCLAttributeDef("device_type", type=t.CharacterString),
+        0x0028: ZCLAttributeDef("feed_back_value", type=t.enum8),
+        0x004B: ZCLAttributeDef("object_id", type=t.FixedList[4, t.uint8_t]),
+        0x004D: ZCLAttributeDef("object_name", type=t.CharacterString),
+        0x004F: ZCLAttributeDef("object_type", type=t.enum16),
+        0x00A8: ZCLAttributeDef("profile_name", type=t.CharacterString),
     }
     server_commands: dict[int, ZCLCommandDef] = {}
     client_commands: dict[int, ZCLCommandDef] = {}
@@ -315,13 +315,13 @@ class MultistateOutputExtended(Cluster):
     cluster_id = 0x0611
     ep_attribute = "bacnet_extended_multistate_output"
     attributes: dict[int, ZCLAttributeDef] = {
-        0x0000: ("acked_transitions", t.bitmap8),
-        0x0011: ("notification_class", t.uint16_t),
-        0x0023: ("event_enable", t.bitmap8),
-        0x0024: ("event_state", t.enum8),
-        0x0048: ("notify_type", t.enum8),
-        0x0071: ("time_delay", t.uint8_t),
-        # 0x0082: ('event_time_stamps', TODO.array),  # Array[3] of (16-bit unsigned
+        0x0000: ZCLAttributeDef("acked_transitions", type=t.bitmap8),
+        0x0011: ZCLAttributeDef("notification_class", type=t.uint16_t),
+        0x0023: ZCLAttributeDef("event_enable", type=t.bitmap8),
+        0x0024: ZCLAttributeDef("event_state", type=t.enum8),
+        0x0048: ZCLAttributeDef("notify_type", type=t.enum8),
+        0x0071: ZCLAttributeDef("time_delay", type=t.uint8_t),
+        # 0x0082: ZCLAttributeDef('event_time_stamps', type=TODO.array),  # Array[3] of (16-bit unsigned
         # integer, time of day, or structure of (date, time of day))
     }
     server_commands: dict[int, ZCLCommandDef] = {}
@@ -332,10 +332,10 @@ class MultistateValueRegular(Cluster):
     cluster_id = 0x0612
     ep_attribute = "bacnet_regular_multistate_value"
     attributes: dict[int, ZCLAttributeDef] = {
-        0x004B: ("object_id", t.FixedList[4, t.uint8_t]),
-        0x004D: ("object_name", t.CharacterString),
-        0x004F: ("object_type", t.enum16),
-        0x00A8: ("profile_name", t.CharacterString),
+        0x004B: ZCLAttributeDef("object_id", type=t.FixedList[4, t.uint8_t]),
+        0x004D: ZCLAttributeDef("object_name", type=t.CharacterString),
+        0x004F: ZCLAttributeDef("object_type", type=t.enum16),
+        0x00A8: ZCLAttributeDef("profile_name", type=t.CharacterString),
     }
     server_commands: dict[int, ZCLCommandDef] = {}
     client_commands: dict[int, ZCLCommandDef] = {}
@@ -345,15 +345,15 @@ class MultistateValueExtended(Cluster):
     cluster_id = 0x0613
     ep_attribute = "bacnet_extended_multistate_value"
     attributes: dict[int, ZCLAttributeDef] = {
-        0x0000: ("acked_transitions", t.bitmap8),
-        0x0006: ("alarm_value", t.uint16_t),
-        0x0011: ("notification_class", t.uint16_t),
-        0x0023: ("event_enable", t.bitmap8),
-        0x0024: ("event_state", t.enum8),
-        0x0025: ("fault_values", t.uint16_t),
-        0x0048: ("notify_type", t.enum8),
-        0x0071: ("time_delay", t.uint8_t),
-        # 0x0082: ('event_time_stamps', TODO.array),  # Array[3] of (16-bit unsigned
+        0x0000: ZCLAttributeDef("acked_transitions", type=t.bitmap8),
+        0x0006: ZCLAttributeDef("alarm_value", type=t.uint16_t),
+        0x0011: ZCLAttributeDef("notification_class", type=t.uint16_t),
+        0x0023: ZCLAttributeDef("event_enable", type=t.bitmap8),
+        0x0024: ZCLAttributeDef("event_state", type=t.enum8),
+        0x0025: ZCLAttributeDef("fault_values", type=t.uint16_t),
+        0x0048: ZCLAttributeDef("notify_type", type=t.enum8),
+        0x0071: ZCLAttributeDef("time_delay", type=t.uint8_t),
+        # 0x0082: ZCLAttributeDef('event_time_stamps', type=TODO.array),  # Array[3] of (16-bit unsigned
         # integer, time of day, or structure of (date, time of day))
     }
     server_commands: dict[int, ZCLCommandDef] = {}
