@@ -270,7 +270,7 @@ class _NeighborEnums:
         On = 0x1
         Unknown = 0x2
 
-    class RelationShip(t.enum3):
+    class Relationship(t.enum3):
         Parent = 0x0
         Child = 0x1
         Sibling = 0x2
@@ -289,7 +289,10 @@ class Neighbor(t.Struct):
     PermitJoins = _NeighborEnums.PermitJoins
     DeviceType = _NeighborEnums.DeviceType
     RxOnWhenIdle = _NeighborEnums.RxOnWhenIdle
-    RelationShip = _NeighborEnums.RelationShip
+    Relationship = _NeighborEnums.Relationship
+
+    # Backwards-compatible alternate spelling
+    RelationShip = Relationship
 
     extended_pan_id: t.ExtendedPanId
     ieee: t.EUI64
@@ -297,7 +300,7 @@ class Neighbor(t.Struct):
 
     device_type: _NeighborEnums.DeviceType
     rx_on_when_idle: _NeighborEnums.RxOnWhenIdle
-    relationship: _NeighborEnums.RelationShip
+    relationship: _NeighborEnums.Relationship
     reserved1: t.uint1_t
 
     permit_joining: _NeighborEnums.PermitJoins
@@ -322,16 +325,41 @@ class Neighbor(t.Struct):
 class Neighbors(t.Struct):
     """Mgmt_Lqi_rsp"""
 
-    entries: t.uint8_t
-    start_index: t.uint8_t
-    neighbor_table_list: t.LVList[Neighbor]
+    Entries: t.uint8_t
+    StartIndex: t.uint8_t
+    NeighborTableList: t.LVList[Neighbor]
+
+
+class RouteStatus(t.enum3):
+    """Route descriptor route status."""
+
+    Active = 0x00
+    Discovery_Underway = 0x01
+    Discovery_Failed = 0x02
+    Inactive = 0x03
+    Validation_Underway = 0x04
+    Reserved_5 = 0x05
+    Reserved_6 = 0x06
+    Reserved_7 = 0x07
 
 
 class Route(t.Struct):
     """Route Descriptor"""
 
     DstNWK: t.NWK
-    RouteStatus: t.uint8_t
+    RouteStatus: RouteStatus
+
+    # Whether the device is a memory constrained concentrator.
+    MemoryConstrained: t.uint1_t
+
+    # The destination is a concentrator that issued a many-to-one request.
+    ManyToOne: t.uint1_t
+
+    # A route record command frame should be sent to the destination prior to the next
+    # data packet.
+    RouteRecordRequired: t.uint1_t
+    Reserved: t.uint2_t
+
     NextHop: t.NWK
 
 
@@ -362,6 +390,11 @@ class Binding(t.Struct):
     SrcEndpoint: t.uint8_t
     ClusterId: t.uint16_t
     DstAddress: MultiAddress
+
+
+class AddrRequestType(t.enum8):
+    Single = 0x00
+    Extended = 0x01
 
 
 class Status(t.enum8):
@@ -500,8 +533,16 @@ class ZDOCmd(t.enum_factory(_CommandID)):
 
 CLUSTERS = {
     # Device and Service Discovery Server Requests
-    ZDOCmd.NWK_addr_req: (IEEE, ("RequestType", t.uint8_t), ("StartIndex", t.uint8_t)),
-    ZDOCmd.IEEE_addr_req: (NWKI, ("RequestType", t.uint8_t), ("StartIndex", t.uint8_t)),
+    ZDOCmd.NWK_addr_req: (
+        IEEE,
+        ("RequestType", AddrRequestType),
+        ("StartIndex", t.uint8_t),
+    ),
+    ZDOCmd.IEEE_addr_req: (
+        NWKI,
+        ("RequestType", AddrRequestType),
+        ("StartIndex", t.uint8_t),
+    ),
     ZDOCmd.Node_Desc_req: (NWKI,),
     ZDOCmd.Power_Desc_req: (NWKI,),
     ZDOCmd.Simple_Desc_req: (NWKI, ("EndPoint", t.uint8_t)),
